JAC.jl Documentation

JAC.AlphaVariation.Settings — Type

struct AlphaVariation.Settings ... defines a type for the details and parameters of computing alpha-variation parameters.

+ calcK                    ::Bool             ... True if the enhancement factor need to be calculated, and false otherwise.
+ printBeforeComputation   ::Bool             ... True if a list of selected levels is printed before the actual computations start. 
+ selectLevels             ::Bool             ... True if individual levels are selected for the computation.
+ selectedLevels           ::Array{Level,1}   ... List of selected levels.

JAC.AutoIonization.Settings — Type

struct Settings ... defines a type for the details and parameters of computing Auger lines.

+ calcAnisotropy          ::Bool                         ... True, if the intrinsic alpha_2,4 angular parameters are to be calculated, 
                                                             and false otherwise.
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).
+ minAugerEnergy          ::Float64                      ... Minimum energy of free (Auger) electrons to be included.
+ maxAugerEnergy          ::Float64                      ... Maximum energy of free (Auger) electrons to be included.
+ maxKappa                ::Int64                        ... Maximum kappa value of partial waves to be included.
+ operator                ::String                       ... Auger operator that is to be used for evaluating the Auger amplitudes: 
                                                             allowed values are: "Coulomb", "Breit", "Coulomb+Breit"

JAC.CoulombExcitation.Settings — Type

struct CoulombExcitation.Settings ... defines a type for the details and parameters of computing Coulomb-excitation lines.

+ multipoles              ::Array{EmMultipole}           ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge}              ... Specifies the gauges to be included into the computations.
+ energies                ::Array{Float64,1}             ... List of ... energies.
+ calcAlignment           ::Bool                         ... True, if alignment parameters to be calculated and false otherwise.
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).

JAC.CoulombIonization.Settings — Type

struct CoulombIonization.Settings ... defines a type for the details and parameters of computing Coulomb-ionization lines.

+ multipoles              ::Array{EmMultipole}           ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge}              ... Specifies the gauges to be included into the computations.
+ energies                ::Array{Float64,1}             ... List of ... energies.
+ calcAlignment           ::Bool                         ... True, if alignment parameters to be calculated and false otherwise.
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).

JAC.Dielectronic.Settings — Type

struct Dielectronic.Settings ... defines a type for the details and parameters of computing dielectronic recombination pathways.

+ multipoles              ::Array{EmMultipoles}   ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge}       ... Specifies the gauges to be included into the computations.
+ printBeforeComputation  ::Bool                  ... True, if all energies and pathways are printed before their evaluation.
+ selectPathways          ::Bool                               ... True if particular pathways are selected for the computations.
+ selectedPathways        ::Array{Tuple{Int64,Int64,Int64},1}  ... List of list of pathways, given by tupels (inital, inmediate, final).
+ electronEnergyShift     ::Float64               ... An overall energy shift for all electron energies (i.e. from the initial to
                                                      the resonance levels.
+ photonEnergyShift       ::Float64               ... An overall energy shift for all photon energies (i.e. from the resonance to
                                                      the final levels.
+ mimimumPhotonEnergy     ::Float64               ... minimum transition energy for which photon transitions are included into the
                                                      evaluation.
+ augerOperator           ::String                ... Auger operator that is to be used for evaluating the Auger amplitudes: 
                                                      allowed values are: "Coulomb", "Breit", "Coulomb+Breit"

JAC.Einstein.Settings — Type

struct Einstein.Settings ... defines a type for the details and parameters of computing Einstein lines and coefficients

+ multipoles              ::Array{EmMultipoles}          ... Specifies the multipoles of the radiation field that are to be included.
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).
+ photonEnergyShift       ::Float64                      ... An overall energy shift for all photon energies.
+ mimimumPhotonEnergy     ::Float64                      ... minimum transition energy for which (photon) transitions are included into the
                                                             computation.
+ maximumPhotonEnergy     ::Float64                      ... maximum transition energy for which (photon) transitions are included.

JAC.FormFactor.Settings — Type

struct FormFactor.Settings ... defines a type for the details and parameters of computing alpha-variation parameters.

+ qList                    ::Array{Float64,1} ... List of q-values in [a.u.]
+ printBeforeComputation   ::Bool             ... True if a list of selected levels is printed before the actual computations start. 
+ selectLevels             ::Bool             ... True if individual levels are selected for the computation.
+ selectedLevels           ::Array{Level,1}   ... List of selected levels.

JAC.Hfs.Settings — Type

struct Settings ... defines a type for the details and parameters of computing HFS A and B coefficients.

+ calcT1                    ::Bool             ... True if T1-amplitudes (HFS A values) need to be calculated, and false otherwise.
+ calcT2                    ::Bool             ... True if T2-amplitudes (HFS B values) need to be calculated, and false otherwise.
+ calcNondiagonal           ::Bool             ... True if also (non-)diagonal hyperfine amplitudes are to be calculated and
                                                   printed, and false otherwise.
+ calcIJFexpansion          ::Bool             ... True if the selected atomic levels are to be represented in a IJF-coupled basis,
                                                   and false otherwise.
+ printBeforeComputation    ::Bool             ... True if a list of selected levels is printed before the actual computations start. 
+ printDeltaEF              ::Bool             ... True if the energy shift of E_F (with regard to E_J) is to be printed, and false otherwise.
+ selectLevels              ::Bool             ... True if individual levels are selected for the computation.
+ selectedLevels            ::Array{Int64,1}   ... List of selected levels.

JAC.ImpactExcitationAutoion.Settings — Type

struct ImpactExcitationAutoion.Settings ... defines a type for the details and parameters of computing electron-impact excitation-autoionization pathways |i(N)> –> |m(N)> –> |f(N-1)>.

+ electronEnergies        ::Array{Float64,1}                   ... List of impact-energies of the incoming elecgtrons.
+ printBeforeComputation  ::Bool                               ... True, if all energies and lines are printed before their evaluation.
+ selectPathways          ::Bool                               ... True if particular pathways are selected for the computations.
+ selectedPathways        ::Array{Tuple{Int64,Int64,Int64},1}  ... List of list of pathways, given by tupels (inital, inmediate, final).
+ maxKappa                ::Int64                              ... Maximum kappa value of partial waves to be included.

JAC.PhotoExcitation.Settings — Type

struct PhotoExcitation.Settings ... defines a type for the details and parameters of computing photo-excitation lines.

+ multipoles              ::Array{EmMultipole,1}         ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge,1}            ... Specifies the gauges to be included into the computations.
+ calcForStokes           ::Bool                         ... True, if the excitation cross sections are to be calculated (and false otherwise)
                                                             for given Stokes parameter of the incident plane-wave photons.
+ calcPhotonDm            ::Bool                         ... True, if the photon density matrix of a subsequently emitted fluorescence photon is to be
                                                             calculated and false otherwise. 
+ calcTensors             ::Bool                         ... True, if the statistical tensors of the excited atom are to be calculated and false otherwise. 
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).
+ photonEnergyShift       ::Float64                      ... An overall energy shift for all photon energies.
+ mimimumPhotonEnergy     ::Float64                      ... minimum transition energy for which (photon) transitions are included into the
                                                             computation.
+ maximumPhotonEnergy     ::Float64                      ... maximum transition energy for which (photon) transitions are included.
+ stokes                  ::ExpStokes                    ... Stokes parameters of the incident radiation.

JAC.PhotoIonization.Settings — Type

struct PhotoIonization.Settings ... defines a type for the details and parameters of computing photoionization lines.

+ multipoles              ::Array{EmMultipole}           ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge}              ... Specifies the gauges to be included into the computations.
+ photonEnergies          ::Array{Float64,1}             ... List of photon energies.  
+ calcAnisotropy          ::Bool                         ... True, if the beta anisotropy parameters are to be calculated and false otherwise (o/w). 
+ calcPartialCs           ::Bool                         ... True, if partial cross sections are to be calculated and false otherwise.  
+ calcTensors             ::Bool                         ... True, if statistical tensors of the excited atom are to be calculated and false o/w. 
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).
+ stokes                  ::ExpStokes                    ... Stokes parameters of the incident radiation.

JAC.PhotoIonizationAutoion.Settings — Type

struct PhotoIonizationAutoion.Settings ... defines a type for the details and parameters of computing photon-impact excitation-autoionization pathways |i(N)> –> |m(N)> –> |f(N-1)>.

+ multipoles              ::Array{EmMultipole,1}               ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge,1}                  ... Specifies the gauges to be included into the computations.
+ printBeforeComputation  ::Bool                               ... True, if all energies and lines are printed before their evaluation.
+ selectPathways          ::Bool                               ... True if particular pathways are selected for the computations.
+ selectedPathways        ::Array{Tuple{Int64,Int64,Int64},1}  ... List of list of pathways, given by tupels (inital, inmediate, final).
+ maxKappa                ::Int64                              ... Maximum kappa value of partial waves to be included.

JAC.RayleighCompton.Settings — Type

struct RayleighCompton.Settings ... defines a type for the settings in estimating lastic Rayleigh and inelastic Compton photon scattering cross sections

+ multipoles              ::Array{EmMultipole}           ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge}              ... Specifies the gauges to be included into the computations.
+ photonEnergies          ::Array{Float64,1}             ... List of photon energies.
+ printBeforeComputation  ::Bool                         ... True, if all energies and lines are printed before their evaluation.
+ selectLines             ::Bool                         ... True, if lines are selected individually for the computations.
+ selectedLines           ::Array{Tuple{Int64,Int64},1}  ... List of lines, given by tupels (inital-level, final-level).

JAC.Atomic.Computation — Type

struct Computation ... defines a type for defining (the model of simple) atomic computation of a single multiplet, including the SCF and CI as well as level properties and transition property calculations.

+ name                           ::String                        ... A name associated to the computation.
+ nuclearModel                   ::Nuclear.Model          ... Model, charge and parameters of the nucleus.
+ grid                           ::Radial.Grid                   ... The radial grid to be used for the computation.
##x + calcLevelProperties            ::Bool                          ... True, if level structures and properties are to be calculated
+ properties                     ::Array{AtomicLevelProperty,1}  ... List of atomic properties to be calculated.
+ configs                        ::Array{Configuration,1}        ... A list of non-relativistic configurations.
+ asfSettings                    ::AsfSettings                   ... Provides the settings for the SCF process and for the CI and QED calculations.
+ initialConfigs                 ::Array{Configuration,1}        ... A list of initial-state configurations for some transition 
                                                                     property calculation, such as radiative transition, Auger, etc. 
+ initialAsfSettings             ::AsfSettings                   ... Provides the SCF and CI settings for the initial-state multiplet.
+ intermediateConfigs            ::Array{Configuration,1}        ... A list of initial-state configurations.
+ intermediateAsfSettings        ::AsfSettings                   ... Provides the SCF settings for the intermediate-state multiplet.
+ finalConfigs                   ::Array{Configuration,1}        ... A list of final-state configurations.
+ finalAsfSettings               ::AsfSettings                   ... Provides the SCF and CI settings for the final-state multiplet.
+ alphaSettings                  ::AlphaVariation.Settings       ... Settings for alpha-variation parameter calculations.
+ einsteinSettings               ::Einstein.Settings             ... Settings for Einstein coefficient calculations.
+ formSettings                   ::FormFactor.Settings           ... Settings for atomic form factor calculations.
##x + greenSettings                  ::GreenFunction.Settings        ... Settings for approximate Green function calculations.
+ hfsSettings                    ::Hfs.Settings                  ... Settings for hyperfine parameter calculations.
+ isotopeSettings                ::IsotopeShift.Settings         ... Settings for isotope shift parameter calculations.
+ plasmaSettings                 ::PlasmaShift.Settings          ... Settings for plasma-shift calculations.
+ polaritySettings               ::MultipolePolarizibility.Settings   ... Settings for plasma-shift calculations.
+ yieldSettings                  ::DecayYield.Settings           ... Settings for fluoresence and Auger yield calculations.
+ zeemanSettings                 ::LandeZeeman.Settings          ... Settings for Lande-Zeeman coefficient calculations.
+ process                        ::JAC.AtomicProcess             ... An (additional) process for which the properties are to be evaluated 
                                                                     for the given initial- and final-state configurations.
+ processSettings                ::Union{JAC.PhotoEmission.Settings, JAC.AutoIonization.Settings, JAC.PlasmaShift.AugerSettings, 
                                         JAC.PhotoIonization.Settings, JAC.PlasmaShift.PhotoSettings, 
                                         JAC.PhotoExcitation.Settings, JAC.PhotoExcitationAutoion.Settings, JAC.PhotoRecombination.Settings, 
                                         JAC.ImpactExcitation.Settings, JAC.Dielectronic.Settings, RadiativeAuger.Settings,
                                         JAC.PairAnnihilation1Photon.Settings, JAC.ImpactExcitationAutoion.Settings, 
                                         JAC.MultiPhotonDeExcitation.Settings, JAC.CoulombExcitation.Settings, 
                                         JAC.CoulombIonization.Settings} ... Provides the settings for the selected process.

JAC.Atomic.Representation — Type

struct Atomic.Representation ... a struct for defining an atomic representation. Such representations often refer to approximate wave function approximations of one or several levels but may concern also a mean-field basis (for some multiplet of some given configurations) or Green functions, etc.

+ name             ::String                      ... to assign a name to the given model.
+ nuclearModel     ::Nuclear.Model               ... Model, charge and parameters of the nucleus.
+ grid             ::Radial.Grid                 ... The radial grid to be used for the computation.
+ refConfigs       ::Array{Configuration,1}      ... List of references configurations, at least 1.
+ repType          ::AbstractRepresentationType  ... Specifies the particular representation.

JAC.Atomic.CiExpansion — Type

struct Atomic.CiExpansion <: AbstractRepresentationType ... a struct to represent (and generate) a configuration-interaction representation.

+ applyOrbitals    ::Dict{Subshell, Orbital}
+ excitations      ::Atomic.RasStep            ... Excitations beyond refConfigs.
+ settings         ::Atomic.CiSettings         ... Settings for the given CI expansion

JAC.Atomic.CiSettings — Type

struct Atomic.CiSettings ... a struct for defining the settings for a configuration-interaction (CI) expansion.

+ breitCI              ::Bool               ... logical flag to include Breit interactions.
+ selectLevelsCI       ::Bool               ... true, if specific level (number)s have been selected.
+ selectedLevelsCI     ::Array{Int64,1}     ... Level number that have been selected.
+ selectSymmetriesCI   ::Bool                          ... true, if specific level symmetries have been selected.
+ selectedSymmetriesCI ::Array{LevelSymmetry,1}        ... Level symmetries that have been selected.

JAC.Atomic.RasExpansion — Type

struct Atomic.RasExpansion <: AbstractRepresentationType ... a struct to represent (and generate) a restricted active-space representation.

+ symmetry         ::LevelSymmetry             ... Symmetry of the levels/CSF in the many-electron basis.
+ NoElectrons      ::Int64                     ... Number of electrons.
+ steps            ::Array{Atomic.RasStep,1}   ... List of SCF steps that are to be done in this model computation.
+ settings         ::Atomic.RasSettings        ... Settings for the given RAS computation

JAC.Atomic.RasSettings — Type

struct Atomic.RasSettings ... a struct for defining the settings for a restricted active-space computations.

+ levelsScf            ::Array{Int64,1}     ... Levels on which the optimization need to be carried out.
+ maxIterationsScf     ::Int64              ... maximum number of SCF iterations in each RAS step.
+ accuracyScf          ::Float64            ... convergence criterion for the SCF field.

+ breitCI              ::Bool               ... logical flag to include Breit interactions.
+ selectLevelsCI       ::Bool               ... true, if specific level (number)s have been selected.
+ selectedLevelsCI     ::Array{Int64,1}     ... Level number that have been selected.

JAC.Atomic.GreenExpansion — Type

struct Atomic.GreenExpansion <: AbstractRepresentationType ... defines a type to keep an (approximate) Green (function) expansion that is associated with a given set of reference configurations.

+ approach          ::Atomic.AbstractGreenApproach    ... Approach used to approximate the representation.
+ excitationScheme  ::Basics.AbstractExcitationScheme ... Applied excitation scheme w.r.t. refConfigs. 
+ levelSymmetries   ::Array{LevelSymmetry,1}          ... Total symmetries J^P to be included into Green expansion.
+ NoElectrons       ::Int64                           ... Number of electrons.
+ settings          ::Atomic.GreenSettings            ... settings for the Green (function) expansion.

JAC.Atomic.GreenSettings — Type

struct Atomic.GreenSettings ... defines a type for defining the details and parameters of the approximate Green (function) expansion.

+ nMax                     ::Int64            ... maximum principal quantum numbers of (single-electron) 
                                                  excitations that are to be included into the representation.
+ lValues                  ::Array{Int64,1}   ... List of (non-relativistic) orbital angular momenta for which
                                                  (single-electron) excitations are to be included.
+ printBeforeComputation   ::Bool             ... True if a short overview is to be printed before. 
+ selectLevels             ::Bool             ... True if individual levels are selected for the computation.
+ selectedLevels           ::Array{Int64,1}   ... List of selected levels.

JAC.Cascade.Computation — Type

struct Cascade.Computation ... defines a type for a cascade computation, i.e. for the computation of a whole excitation and/or decay cascade. The data from this computation can be modified, adapted and refined to the practical needs before the actual computations are carried out. Initially, this struct contains the physical metadata about the cascade to be calculated but gets enlarged in course of the computation to keep also wave functions, level multiplets, etc.

+ name               ::String                         ... A name for the cascade
+ nuclearModel       ::Nuclear.Model                  ... Model, charge and parameters of the nucleus.
+ grid               ::Radial.Grid                    ... The radial grid to be used for the computation.
+ asfSettings        ::AsfSettings                    ... Provides the settings for the SCF process.
+ approach           ::Cascade.AbstractApproach       ... Computational approach/model that is applied to generate and evaluate the 
                                                          cascade; possible approaches are: {'single-configuration', ...}
+ processes          ::Array{Basics.AtomicProcess,1}   ... List of the atomic processes that are supported and should be included into the 
                                                          cascade.
+ initialConfs       ::Array{Configuration,1}         ... List of one or several configurations that contain the level(s) from which the 
                                                          cascade starts.
+ initialLevels      ::Array{Tuple{Int64,Float64},1}  ... List of one or several (tupels of) levels together with their relative population 
                                                          from which the cascade starts.
+ maxElectronLoss    ::Int64                          ... (Maximum) Number of electrons in which the initial- and final-state 
                                                          configurations can differ from each other; this also determines the maximal steps 
                                                          of any particular decay path.
+ NoShakeDisplacements ::Int64                        ... Maximum number of electron displacements due to shake-up or shake-down processes 
                                                          in any individual step of the cascade.
+ shakeFromShells ::Array{Shell,1}                    ... List of shells from which shake transitions may occur.
+ shakeToShells   ::Array{Shell,1}                    ... List of shells into which shake transitions may occur.
+ steps           ::Array{Cascade.Step,1}             ... List of individual steps between well-defined atomic multiplets that are 
                                                          included into the cascade.

JAC.Cascade.SimulationSettings — Type

struct Cascade.SimulationSettings ... defines settings for performing the simulation of some cascade (data).

+ minElectronEnergy ::Float64     ... Minimum electron energy for the simulation of electron spectra.
+ maxElectronEnergy ::Float64     ... Maximum electron energy for the simulation of electron spectra.
+ minPhotonEnergy   ::Float64     ... Minimum photon energy for the simulation of electron spectra.
+ maxPhotonEnergy   ::Float64     ... Maximum photon energy for the simulation of electron spectra..

JAC.Cascade.Simulation — Type

struct Cascade.Simulation ... defines a simulation on some given cascade (data).

+ properties      ::Array{Cascade.Property,1}   ... Properties that are considered in this simulation of the cascade (data).
+ method          ::Cascade.SimulationMethod    ... Method that is used in the cascade simulation; cf. Cascade.SimulationMethod.
+ settings        ::Cascade.SimulationSettings  ... Settings for performing these simulations.

JAC.Basics.add — Function

Basics.add(ma::AngularM64, mb::AngularM64) ... adds the projections of the angular momenta ma + mb and returns a mc::AngularM64.

Basics.add(pota::Radial.Potential, potb::Radial.Potential) ... to add two radial potentials together that are defined on the same grid. A potential::RadialPotential is returned that inherits its radial size from the potential that is defined in a larger range of r-values.

JAC.Basics.compute — Function

Basics.compute("angular coefficients: e-e, Ratip2013", csfa::CsfR, csfb::CsfR) ... to compute the angular coefficients in the decomposition of a (reduced) many-electron matrix element with a general rank-0 electron-electron interaction operator $⟨csf_a ||V(e-e)|| csfb⟩ = \sum_t T(a_t, b_t, c_t, d_t) * R^k (a_t, b_t, c_t, d_t)$ by a call to the Fortran procedure anco_calculate_csf_pair of the RATIP program; a Tuple{Array{AngularTcoeff,1},Array{AngularVcoeff,1}}` is returned.

Basics.compute("angular coefficients: 1-p, Ratip2013", rank::Int64, csfa::CsfR, csfb::CsfR) ... to compute the the angular coefficients in the decomposition of a (reduced) many-electron matrix element with a general single-particle operator of the given rank $⟨csf_a ||O^rank|| csfb⟩ = \sum_t T(a_t, b_t) * R (a_t, b_t)$ by a call to the Fortran procedure anco_calculate_csf_pair_1p of the RATIP program; an Array{AngularTcoeff,1} is returned.

Basics.compute("angular coefficients: 1-p, Grasp92", parity, rank::Integer, csfa::CsfR, csfb::CsfR) ... to compute the the angular coefficients in the decomposition of a (reduced) many-electron matrix element with a general single-particle operator of the given parity and rank by a call to the Fortran procedure mct_generate_coefficients of the RATIP program; an Array{AngularTcoeff,1} is returned.

Basics.compute("matrix: CI, J^P symmetry", JP::LevelSymmetry, basis::Basis, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings; printout::Bool=true) ... to compute the CI matrix for a given J^P symmetry block of basis and by making use of the nuclear model and the grid; a matrix::Array{Float64,2} is returned.

Basics.compute("matrix: CI for plasma, J^P symmetry", JP::LevelSymmetry, basis::Basis, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings, plasmaSettings::PlasmaShift.Settings; printout::Bool=true) ... to compute the CI matrix for a given J^P symmetry block of basis and by making use of the nuclear model and the grid; a matrix::Array{Float64,2} is returned.

Basics.compute("radial orbital: NR, Bunge (1993)", subshell::Subshell, Z::Int64) ... to compute a radial orbital::Orbital for the given subshell and nuclear charge by using the Roothan-Hartree-Fock data by Bunge et al., Atomic Data and Nuclear Data Tables 53 (1993) 113, as obtained for a non-relativistic RHF computation of the neutral atom. These functions are used a large component, and the small component is obtained from the kinetic balance condition. Radial orbitals can be obtained for the ground-state configuration for all elements with 2 <= Z <= 54. Basics.compute("radial orbital: NR, McLean (1981)", subshell::Subshell, Z::Int64) ... to compute a radial orbital::Orbital for the given subshell and nuclear charge by using the Roothan-Hartree-Fock data by McLean and McLean, Atomic Data and Nuclear Data Tables 26 (1981) 197., as obtained for a non-relativistic RHF computation of the neutral atom. These functions are used a large component, and the small component is obtained from the kinetic balance condition. Radial orbitals can be obtained for the ground-state configuration for all elements with 55 <= Z <= 92.

Basics.compute("radial potential: core-Hartree", grid::Radial.Grid, level::Level) ... to compute a (radial) core-Hartree potential from the given list of orbitals; cf. Basics.computePotentialCoreHartree. A potential::RadialPotential is returned.

Basics.compute("radial potential: Kohn-Sham", grid::Radial.Grid, level::Level) ... to compute a (radial) Kohn-Sham potential from the given list of orbitals; cf. Basics.computePotentialKohnSham. A potential::RadialPotential is returned.

Basics.compute("radial potential: Dirac-Fock-Slater", grid::Radial.Grid, level::Level) ... to compute a (radial) Dirac-Fock-Slater potential from the given list of orbitals; this potential is rather simple but includes some undesired self-interaction and exhibits an asymptotic behaviour. Cf. Basics.computePotentialDFS. A potential::RadialPotential is returned.

Basics.compute("radial potential: Dirac-Fock-Slater", grid::Radial.Grid, basis::Basis) ... to compute a (radial) Dirac-Fock-Slater potential but for mean occupation of the given basis. Cf. Basics.computePotentialDFS. A potential::RadialPotential is returned.

Basics.compute("radial potential: extended Hartree", grid::Radial.Grid, level::Level) ... to compute a (radial) extended Hartree potential from the given list of orbitals; it is a local potential that is based on direct and exchange coefficients as obtained from the configuration-averaged energy. Cf. Basics.computePotentialExtendedHartree. A potential::RadialPotential is returned.

JAC.Basics.display — Function

Basics.display("constants") or ("physical constants") ... to display (all) currently defined physical constants; nothing is returned if not indicated otherwise. Cf. Defaults.setDefaults().

Basics.display("settings") ... to display (all) currently defined settings of the JAC module.

Basics.display(stream::IO, configs::Array{Configuration,1}) ... displays the generated list of configurations in a compact form; nothing is returned in this case.

Basics.display(stream::IO, orbitals::Dict{Subshell, Orbital}) ... displays the (generated) orbitals in a neat and compact form; nothing is returned in this case.

Basics.display(stream::IO, channels::Array{Atomic.GreenChannel,1}) ... displays the (generated Green function) channels in a neat and compact form; nothing is returned in this case.

JAC.Basics.generate — Function

Basics.generate(representation::Atomic.Representation) ... to generate an atomic representation as specified by the representation.repType::AbstractRepresentationType. All relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.perform(representation::Atomic.Representation; output=true) ... to generate the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the representation but can easily accessed by the keys of this dictionary.

Basics.generate(repType::Atomic.MeanFieldBasis, representation::Atomic.Representation) ... to generate a mean-field basis (representation) for a set of reference configurations; all relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.generate(repType::Atomic.MeanFieldBasis, representation::Atomic.Representation; output=true) ... to generate the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the representation but can easily accessed by the keys of this dictionary.

Basics.generate(repType::Atomic.CiExpansion, representation::Atomic.Representation) ... to generate a configuration-interaction expansion for a single level symmetry and based on a set of reference configurations and a number of pre-specified steps. All relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.generate(repType::Atomic.CiExpansion, representation::Atomic.Representation; output=true) ... to generate the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the computations but can easily accessed by the keys of this dictionary.

Basics.generate(repType::Atomic.RasExpansion, representation::Atomic.Representation) ... to generate a restricted active-space expansion for a single level symmetry and based on a set of reference configurations and a number of pre-specified steps. All relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.generate(repType::Atomic.RasExpansion, representation::Atomic.Representation; output=true) ... to generate the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the computations but can easily accessed by the keys of this dictionary.

Basics.generate(repType::Atomic.GreenExpansion, representation::Atomic.Representation) ... to generate a Green (function) expansion for a given approach and excitation scheme of the electron, based on a set of reference configurations, a list of level symmetries as well as for given settings. All relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.generate(repType::Atomic.GreenExpansion, representation::Atomic.Representation; output=true) ... to generate the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the representation but can easily accessed by the keys of this dictionary.

Basics.generate("condensed multiplet: by single weight", multiplet::Multiplet) ... to condense/reduce the number of CSF in the basis of the given multiplet due to a single 'weight'; a multiplet::Multiplet is returned. Not yet implemented !

Basics.generate("configuration list: NR, from basis", basis::Basis) ... to (re-) generate the list of NR configurations from the given basis; a confList::Array{Configuration,1} is returned.

Basics.generate("configuration list: NR, single-configuration", refConf::Configuration, NoExcitations::Int64, fromShells::Array{Shell,1}, toShells::Array{Shell,1}) ... to generate a non-relativistic configuration list, including the given reference configuration (refConf) and with all configurations that differ by NoExcitations from the fromShells into the toShells; an Array{Configuration,1} is returned.

Basics.generate("configuration list: relativistic", conf::Configuration) ... to split/decompose a non-relativistic configuration into an list of relativistic ConfigurationR[]. The proper occupuation of the relativistic subshells is taken into account.

Basics.generate("CSF list: from single ConfigurationR", conf::ConfigurationR, subshellList::Array{Subshell,1}) ... to construct from a given (relativistic) configuration all possible CSF with regard to the subshell order as specified by subshellList; a list::Array{CsfR,1} is returned.

Basics.generate("shells: ordered list for NR configurations", confs::Array{Configuration,1}) ... to generate for confs, i.e. all the given (non-relativistic) configurations, a common and ordered shell list; a list::Array{Shell,1} is returned.

Basics.generate("subshells: ordered list for relativistic configurations", confs::Array{ConfigurationR,1}) ... to generate for confs, i.e. all the given (relativistic) configurations, common and ordered subshell list; a list::Array{Subshell,1} is returned.

Basics.generate("subshells: ordered list for two bases", basisA::Basis, basisB::Basis) ... to generate common and ordered subshell list for the two basis A and B; a list::Array{Subshell,1} is returned.

Basics.generate("single-electron spectrum: STO", N::Int64, potential::Radial.Potential, grid::Radial.Grid; N_0::Int64=30, alpha_0::Float64=1.0, beta_0::Float64=1.1) ... to generate a complete one-electron spectrum with N positive and N negative states, and by using even-tempered Slater-type orbitals (STO) with parameters $lpha_i = lpha_0 eta_0^i$; a spectrum::SingleElecSpectrum is returned where just N0 positive and N_0 negative are kept for later use. Not yet implemented !

Basics.generate("single-electron spectrum: STO, positive", N::Int64, potential::Radial.Potential, grid::Radial.Grid; N_0::Int64=30, alpha_0::Float64=1.0, beta_0::Float64=1.1) ... to generate the same but to return only the N_0 positive states. Not yet implemented !

JAC.Basics.perform — Function

Basics.perform(computation::Atomic.Computation) ... to perform the computation as prescribed by comp. All relevant intermediate and final results are printed to screen (stdout). Nothing is returned.

Basics.perform(computation::Atomic.Computation; output=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the computations but can easily accessed by the keys of this dictionary.

Basics.perform(comp::Cascade.Computation) ... to set-up and perform a cascade computation that starts from a given set of initial configurations and proceeds via various steps until a given number of electrons has been removed or the decay stops at some stable levels with regard to the given atomic processes. The results of all individual steps are printed to screen but nothing is returned otherwise.

Basics.perform(comp::Cascade.Computation; output=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the cascade but can easily accessed by the keys of this dictionary.

Basics.perform(simulation::Cascade.Simulation, data::Cascade.Data) ... to simulate a cascade decay (and excitation) by using the given data::Cascade.Data. Different computational methods and different properties of the ionic system, such as the ion distribution or final-level distribution can be derived and displayed from these simulations. Of course, the details of these simulations strongly depend on the atomic processes and data that have been generated before by performing a computation::Cascade.Computation. The results of all individual steps are printed to screen but nothing is returned otherwise.

Basics.perform(simulation::Cascade.Simulation, data::Cascade.Data; output=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the method and specifications of the cascade but can easily accessed by the keys of this dictionary.

Basics.perform("computation: SCF", configs::Array{Configuration,1}, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings; printout::Bool=true) ... to generate an atomic basis and to compute the self-consistent field (SCF) for this basis due to the given settings; a basis::Basis is returned.

Basics.perform("computation: mutiplet from orbitals, no CI, CSF diagonal", configs::Array{Configuration,1}, initalOrbitals::Dict{Subshell, Orbital}, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings; printout::Bool=true) ... to generate from the given initial orbitals a multiplet of single-CSF levels by just using the diagonal part of the Hamiltonian matrix; a multiplet::Multiplet is returned.

Basics.perform("computation: CI", basis::Basis, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings; printout::Bool=true) ... to set-up and diagonalize from the (SCF) basis the configuration-interaction matrix and to derive and display the level structure of the corresponding multiplet due to the given settings; a multiplet::Multiplet is returned.

Basics.perform("computation: CI for plasma", basis::Basis, nuclearModel::Nuclear.Model, grid::Radial.Grid, settings::AsfSettings, plasmaSettings::PlasmaShift.Settings; printout::Bool=true) ... to set-up and diagonalize from the given (SCF) basis the configuration-interaction matrix and to derive and display the level structure of the corresponding multiplet due to the given settings. Here, the CI matrix includes the modifications of the Hamiltonian due to the given plasmaSettings; a multiplet::Multiplet is returned.

JAC.Basics.tabulate — Function

Basics.tabulate(sa::String, multiplet::Multiplet; stream::IO=stdout) ... tabulates the energies from the multiplet due to different criteria.

("multiplet: energies", multiplet::Multiplet; stream::IO=stdout) ... to tabulate the energies of all levels of the given multiplet into a neat format; nothing is returned.

("multiplet: energy relative to immediately lower level", multiplet::Multiplet; stream::IO=stdout) ... to tabulate the energy splitting between neighboured levels of all levels of the given multiplet into a neat format; nothing is returned.

("multiplet: energy of each level relative to lowest level", multiplet::Multiplet; stream::IO=stdout) ... to tabulate the energy splitting of all levels with regard to the lowest level of the given multiplet into a neat format; nothing is returned.

Basics.tabulate(sa::String, multiplet::Hfs.IJF_Multiplet; stream::IO=stdout) ... tabulates the energies from the multiplet due to different criteria.

("multiplet: energies", multiplet::Hfs.IJF_Multiplet; stream::IO=stdout) ... to tabulate the energies of all hyperfine levels of the given multiplet into a neat format; nothing is returned.

("multiplet: energy of each level relative to lowest level", multiplet::Hfs.IJF_Multiplet; stream::IO=stdout) ... to tabulate the energy splitting of all levels with regard to the lowest level of the given multiplet into a neat format; nothing is returned.

JAC.Defaults.getDefaults — Function

Defaults.getDefaults() ... gives/supplies different information about the (present) framework of the computation or about some given data; cf. Defaults.setDefaults().

("alpha") or ("fine-structure constant alpha") ... to get the (current) value::Float64 of the fine-structure constant alpha.

("electron mass: kg") or ("electron mass: amu") ... to get the (current) value::Float64 of the electron mass in the specified unit.

("framework") ... to give the (current) setting::String of the overall framework.

("electron rest energy") or ("mc^2") ... to get the electron rest energy.

("electron g-factor") ... to give the electron g-factor g_s = 2.00232.

("unit: energy") or ("unit: cross section") or ("unit: rate") ... to get the corresponding (user-defined) unit::String for the current computations.

("standard grid") ... to get the (current standard) grid::Array{Float64,1} to which all radial orbital functions usually refer.

("speed of light: c") ... to get the speed of light in atomic units.

("summary flag/stream") ... to get the logical flag and stream for printing a summary file; a tupel (flag, iostream) is returned.

("ordered shell list: non-relativistic", n_max::Int64) ... to give an ordered list of non-relativistic shells::Array{Shell,1} up to the (maximum) principal number n_max.

("ordered subshell list: relativistic", n_max::Int64) ... to give an ordered list of relativistic subshells::Array{Subshell,1} up to the (maximum) principal number n_max.

JAC.Defaults.setDefaults — Function

Defaults.setDefaults() ... (re-) defines some 'standard' settings which are common to all the computations with the JAC module, and which can be 'overwritten' by the user. –- An improper setting of some variable may lead to an error message, if recognized immediately. The following defaults apply if not specified otherwise by the user: the framework is 'relativistic', energies are given in eV and cross sections in barn. Note that, internally, atomic units are used throughout for all the computations within the program. nothing is returned if not indicated otherwise.

("framework: relativistic") or ("framework: non-relativistic") ... to define a relativistic or non-relativistic framework for all subsequent computations.

("method: continuum, spherical Bessel") or ("method: continuum, pure sine") or ("method: continuum, asymptotic Coulomb") or ("method: continuum, nonrelativistic Coulomb") or ("method: continuum, Galerkin") ... to define a a method for the generation of the continuum orbitals as (pure) spherical Bessel, pure sine, asymptotic Coulomb, nonrelativistic Coulomb orbital or by means of the B-spline-Galerkin method.

("method: normalization, pure sine") or ("method: normalization, pure Coulomb") ... to define a method for the normalization of the continuum orbitals as asymptotically (pure) sine or Coulomb functions.

("QED model: Petersburg") or ("QED model: Sydney") ... to define a model for the computation of the QED corrections following the work by Shabaev et al. (2011; Petersburg) or Flambaum and Ginges (2004; Syney).

("unit: energy", "eV") or ("unit: energy", "Kayser") or ("unit: energy", "Hartree") or ("unit: energy", "Hz") or ("unit: energy", "Hz") ... to (pre-) define the energy units for all further printouts and communications with the JAC module.

("unit: cross section", "a.u.") or ("unit: cross section", "barn") or ("unit: cross section", "Mbarn") ... to (pre-) define the unit for the printout of cross sections.

("unit: rate", "a.u.") or ("unit: rate", "1/s") ... to (pre-) define the unit for the printout of rates.

("unit: time", "a.u.") or ("unit: time", "sec") or ("unit: time", "fs") or ("unit: time", "as") ... to (pre-) define the unit for the printout and communications of times with the JAC module.

("relativistic subshell list", subshells::Array{Subshell,1}; printout::Bool=true) ... to (pre-) define internally the standard relativistic subshell list on which the standard order of orbitals is based.

("standard grid", grid::Radial.Grid; printout::Bool=true) ... to (pre-) define internally the standard radial grid which is used to represent most orbitals.

("QED: damped-hydrogenic", Znuc::Float64, wa::Array{Float64,1}) ... to (re-) define the lambda-C damped overlap integrals of the lowest kappa-orbitals [ wa1s1/2, wa2p1/2, wa2p3/2, wa3d3/2, wa3d5/2 ] for the (new) nuclear charge Znuc; nothing is returned.