Angular momentum

JAC.AngularMomentum.CL_reduced_me — Method

AngularMomentum.CL_reduced_me(suba::Subshell, L::Int64, subb::Subshell) ... calculates the reduced matrix element of the C^L spherical tensor <suba || C^(L) || subb>; a value::Float64 is returned.

JAC.AngularMomentum.CL_reduced_me_rb — Method

AngularMomentum.CL_reduced_me_rb(suba::Subshell, L::Int64, subb::Subshell) ... calculates the reduced matrix element of the C^L spherical tensor <suba || C^(L) || subb>; a value::Float64 is returned.

JAC.AngularMomentum.CL_reduced_me_sms — Method

AngularMomentum.CL_reduced_me_sms(suba::Subshell, L::Int64, subb::Subshell) ... calculates the reduced matrix element of the C^L spherical tensor <suba || C^(L) || subb> for the specific MS (SMS); a value::Float64 is returned.

JAC.AngularMomentum.ChengI — Method

`AngularMomentum.ChengI

(kapa::Int64, ma::AngularM64, kapb::Int64, mb::AngularM64, L::AngularJ64, M::AngularM64)` ... evaluates the angular I (kappa m, kappa' m', LM) integral as defined by Cheng, NATO summerschool (198x), Eq. (A4.5), and including the full magnetic (orientational) dependence. A value::Float64 is returned.

JAC.AngularMomentum.ChengI — Method

(kapa::Int64, kapb::Int64, L::AngularJ64)` ... evaluates the same angular I (kappa m, kappa' m', LM) integral but without the magnetic (orientational) dependence. A value::Float64 is returned.

JAC.AngularMomentum.ClebschGordan — Method

AngularMomentum.ClebschGordan(ja, ma, jb, mb, Jab, Mab) ... calculates the Clebsch-Gordan coefficient <ja, ma, jb, mb; Jab, Mab> for given quantum numbers by a proper call to a Wigner 3-j symbol. A value::Float64 is returned.

JAC.AngularMomentum.ClebschGordan_old — Method

AngularMomentum.ClebschGordan_old(ja::AngularJ64, ma::AngularM64, jb::AngularJ64, mb::AngularM64, Jab::AngularJ64, Mab::AngularM64) ... calculates the Clebsch-Gordan coefficient <ja, ma, jb, mb; Jab, Mab> for given quantum numbers by a proper call to a Wigner 3-j symbol; a value::Float64 is returned.

JAC.AngularMomentum.JohnsonI — Method

AngularMomentum.JohnsonI(kapa::Int64, ma::AngularM64, kapb::Int64, mb::AngularM64, L::AngularJ64, M::AngularM64) ... evaluates the angular CL (kappa m, kappa' m', L M) integral as defined in his book. A value::Float64 is returned.

JAC.AngularMomentum.Wigner_3j — Method

AngularMomentum.Wigner_3j(a, b, c, m_a, m_b, m_c) ... calculates the value of a Wigner 3-j symbol for given quantum numbers as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); it calls the corresponding function from the GNU Scientific Library. A value::Float64 is returned.

JAC.AngularMomentum.Wigner_3j_old — Method

AngularMomentum.Wigner_3j_old(a::AngularJ64, b::AngularJ64, c::AngularJ64, m_a::AngularM64, m_b::AngularM64, m_c::AngularM64) ... calculates the value of a Wigner 3-j symbol for given quantum numbers by its algebraic formulae as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); a value::Float64 is returned.

JAC.AngularMomentum.Wigner_6j — Method

AngularMomentum.Wigner_6j(a, b, c, d, e, f) ... calculates the value of a Wigner 6-j symbol for given quantum numbers as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); it calls the corresponding function from the GNU Scientific Library. A value::Float64 is returned.

JAC.AngularMomentum.Wigner_6j_old — Method

AngularMomentum.Wigner_6j_old(a::AngularJ64, b::AngularJ64, c::AngularJ64, d::AngularJ64, e::AngularJ64, f::AngularJ64) ... calculates the value of a Wigner 6-j symbol for given quantum numbers by its algebraic formulae as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); a value::Float64 is returned.

JAC.AngularMomentum.Wigner_9j — Method

AngularMomentum.Wigner_9j(a, b, c, d, e, f, g, h, i) ... calculates the value of a Wigner 3-j symbol for given quantum numbers as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); it calls the corresponding function from the GNU Scientific Library. A value::Float64 is returned.

JAC.AngularMomentum.Wigner_9j_old — Method

AngularMomentum.Wigner_9j_old(a::AngularJ64, b::AngularJ64, c::AngularJ64, d::AngularJ64, e::AngularJ64, f::AngularJ64, g::AngularJ64, h::AngularJ64, i::AngularJ64) ... calculates the value of a Wigner 6-j symbol for given quantum numbers by its algebraic formulae as displayed in many texts on the theory of angular momentum (see R. D. Cowan, The Theory of Atomic Structure and Spectra; University of California Press, 1981, p. 142); a value::Float64 is returned.

JAC.AngularMomentum.Wigner_DFunction — Method

AngularMomentum.Wigner_DFunction(j, p, q, alpha::Float64, beta::Float64, gamma::Float64) ... calculates the value of a Wigner D^j_pq (alpha, beta, gamma) for given quantum numbers and (Euler) angles (alpha, beta, gamma). It makes use of the small Wigner d(beta) matrix as the key part. A value::Float64 is returned.

JAC.AngularMomentum.Wigner_dmatrix — Method

AngularMomentum.Wigner_dmatrix(j, mp, m, beta::Float64) ... calculates the value of the small Wigner d^j_m',m (beta) for given quantum numbers and the angle beta. Wigner's formula is applied to evaluate the small Wigner matrix; a value::Float64 is returned.

JAC.AngularMomentum.allowedDoubleKappaCouplingSequence — Method

AngularMomentum.allowedDoubleKappaCouplingSequence(syma::LevelSymmetry, symb::LevelSymmetry, maxKappa::Int64) ... to determine all allowed coupling sequences that fulfill syma + (kappa1, symx, kappa2) –> symb == syma + kappa1 –> symx + kappa2 –> symb, and where kappa1|, |kappa2| <= maxKappa. A list::Array{Tuple{Int64,LevelSymmetry,Int64},1}) of (kappa1, symx, kappa2) is returned.

JAC.AngularMomentum.allowedDoubleKappaSymmetries — Method

AngularMomentum.allowedDoubleKappaSymmetries(syma::LevelSymmetry, kappa1::Int64, kappa2::Int64, symb::LevelSymmetry) ... to determine all allowed level symmetries symx that can be coupled to the sequence syma + kappa1 –> {symx} + kappa2 –> symb. A list::Array{LevelSymmetry,1} of symx is returned.

JAC.AngularMomentum.allowedDoubleKappas — Method

AngularMomentum.allowedDoubleKappas(syma::LevelSymmetry, symb::LevelSymmetry, maxKappa::Int64) ... to determine all allowed pairs of kappa that fulfill syma + (kappa1, kappa2) –> symb, and where |kappa1|, |kappa2| <= maxKappa. A list::Array{Tuple{Int64,Int64},1}) of (kappa1, kappa2) is returned.

JAC.AngularMomentum.allowedKappaSymmetries — Method

AngularMomentum.allowedKappaSymmetries(syma::LevelSymmetry, symb::LevelSymmetry) ... to determine all allowed single-electron symmetries/partial waves kappa (l,j) that can be coupled to the given level symmetries. A list::Array{Int64,1} of kappa-values is returned.

JAC.AngularMomentum.allowedMultipoleSymmetries — Method

AngularMomentum.allowedMultipoleSymmetries(syma::LevelSymmetry, multipole::EmMultipole) ... to determine all allowed level symmetries for which the given multipole can give rise to a non-zero (transition) amplitude; a symList::Array{LevelSymmetry,1} is returned.

JAC.AngularMomentum.allowedTotalSymmetries — Method

AngularMomentum.allowedTotalSymmetries(symf::LevelSymmetry, mp2::EmMultipole, mp1::EmMultipole, symi::LevelSymmetry) ... to determine all allowed total symmetries J^P that can be constructed by coupling a multipole wave mp1 to the initial symmetry symi, and which can be further coupled with mp2 to the final symmetry symf. A list::Array{LevelSymmetry,1} of total symmetries is returned.

JAC.AngularMomentum.allowedTotalSymmetries — Method

AngularMomentum.allowedTotalSymmetries(syma::LevelSymmetry, kappa::Int64) ... to determine all allowed total symmetries J^P that can be constructed by coupling a partial wave kappa (l,j) to the given level symmetry syma. A list::Array{LevelSymmetry,1} of total symmetries is returned.

JAC.AngularMomentum.bracket — Method

AngularMomentum.bracket(jList::Array{AngularJ64,1}) ... to compute the bracket [a, b, c, ... ] = (2a+1) * (2b+1) * (2b+1) * ... of the given angular momenta. A value::Int64 is returned.

JAC.AngularMomentum.isAllowedMultipole — Method

AngularMomentum.isAllowedMultipole(syma::LevelSymmetry, multipole::EmMultipole, symb::LevelSymmetry) ... evaluates to true if the given multipole may connect the two level symmetries, and false otherwise.

JAC.AngularMomentum.isTriangle — Method

AngularMomentum.isTriangle(ja::AngularJ64, jb::AngularJ64, jc::AngularJ64) ... evaluates to true if Delta(ja,jb,jc) = 1, ie. if the angular momenta ja, jb and jc can couple to each other, and false otherwise.

JAC.AngularMomentum.isTriangle — Method

AngularMomentum.isTriangle(ja::Int64, jb::Int64, jc::Int64) ... evaluates to true if Delta(ja,jb,jc) = 1, ie. if the three integer (length) ja, jb and jc can form a triangle, and false otherwise.

JAC.AngularMomentum.kappa_j — Method

AngularMomentum.kappa_j(kappa::Int64) ... calculates the j::AngularJ64 value of a given kappa.

JAC.AngularMomentum.kappa_l — Method

AngularMomentum.kappa_l(kappa::Int64) ... calculates the l::AngularJ64 value of a given kappa.

JAC.AngularMomentum.m_values — Method

AngularMomentum.m_values(j::AngularJ64) ... returns a list of m-values for given j::AngularJ64.

JAC.AngularMomentum.oneJ — Method

AngularMomentum.oneJ(ja::AngularJ64) ... calculates ja; a (positive) value::Float64 is returned.

JAC.AngularMomentum.parityEmMultipolePi — Method

AngularMomentum.parityEmMultipolePi(pa::Parity, multipole::EmMultipole, pb::Parity) ... evaluates to true if the given multipole fullfills the parity selection rule pi(a, multipole, b) = 1, and false otherwise. This includes a proper test for both, electric and magnetic multipoles, based on multipole.electric.

JAC.AngularMomentum.phaseFactor — Method

AngularMomentum.phaseFactor(list::Array{Any,1}) ... checks and calculates the phase factor (-1)^(ja + mb -jc ...) that occur frequently in angular momentum theory; a value +1. or -1. is returned. Use phaseFactor([ja::Union{AngularJ64,AngularM64), -1, mb::Union{AngularJ64,AngularM64), ..., 1, jc::Union{AngularJ64,AngularM64)]) to specify the phase.

JAC.AngularMomentum.phaseMultipole — Method

AngularMomentum.phaseMultipole(x::ComplexF64, mp::EmMultipole) ... calculates (x)^p with mp = (L,p) and p = 0 (magnetic), p = 1 (electric). A wa ::ComplexF64 is returned.

JAC.AngularMomentum.sigma_reduced_me — Method

AngularMomentum.sigma_reduced_me(suba::Subshell, subb::Subshell) ... calculates the reduced matrix element of the sigma^(1) spherical tensor <suba || sigma^(1) || subb>; a value::Float64 is returned.

JAC.AngularMomentum.sphericalYlm — Method

AngularMomentum.sphericalYlm(l::Int64, m::Int64, theta::Float64, phi::Float64) ... calculates the spherical harmonics for low l-values explicitly. A value::Complex{Float64} is returned. Ylm = sqrt( (2l+1) / (twotwo*pi) ) * spherical_Clm(l,m,theta,phi).

JAC.AngularMomentum.triangularDelta — Method

AngularMomentum.triangularDelta(ja::AngularJ64, jb::AngularJ64, jc::AngularJ64) ... calculates the tringular Delta(ja,jb,jc). The result is 0 if the triangular condition failes and 1 otherwise.

JAC.AngularMomentum.triangularDelta — Method

AngularMomentum.triangularDelta(ia2::Int64, ib2::Int64, ic2::Int64) ... calculates the tringular Delta(ja,jb,jc). The arguments in this integer function are i2a = 2*ja+1, ... The result is 0 if the triangular condition failes and 1 otherwise.

Atomic interaction strength

JAC.InteractionStrength.XLCoefficient — Type

struct InteractionStrength.XLCoefficient ... defines a type for coefficients of the two-electron (Breit) interaction

+ kind      ::Char       ... Kind of integral, either 'S' or 'T'
+ nu        ::Int64      ... Rank of the integral.
+ a         ::Orbital    ... Orbitals a, b, c, d.
+ b         ::Orbital
+ c         ::Orbital
+ d         ::Orbital
+ coeff     ::Float64    ... corresponding coefficient.

JAC.InteractionStrength.MabEmissionJohnsony — Method

InteractionStrength.MabEmissionJohnsony(mp::EmMultipole, gauge::EmGauge, omega::Float64, a::Orbital, b::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <a || O^(Mp,emission) || b> for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned. This procedure has been re-worked due to the book by Johnson (2007).

JAC.InteractionStrength.MbaAbsorptionCheng — Method

InteractionStrength.MbaAbsorptionCheng(mp::EmMultipole, gauge::EmGauge, omega::Float64, b::Orbital, a::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <b || O^(Mp, absorption) || a> for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned.

JAC.InteractionStrength.MbaEmissionAndreyOld — Method

InteractionStrength.MbaEmissionAndrey(mp::EmMultipole, gauge::EmGauge, omega::Float64, b::Orbital, a::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <b || O^(Mp, emission) || a> for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned. At present, only the magnetic matrix elements are implemented. This procedure has been worked out with Andrey but is currently not in use.

JAC.InteractionStrength.MbaEmissionCheng — Method

InteractionStrength.MbaEmissionCheng(mp::EmMultipole, gauge::EmGauge, omega::Float64, b::Orbital, a::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <b || O^(Mp,emission) || a> for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned. This procedure has been first worked out with Andrey; in this case, however, the phases are not under good control, and this gives rise to wrong amplitudes and rates. The procedure is currently not in use.

JAC.InteractionStrength.MbaEmissionJohnsonx — Method

InteractionStrength.MbaEmissionJohnsonx(mp::EmMultipole, gauge::EmGauge, omega::Float64, b::Orbital, a::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <b || O^(Mp,emission) || a> for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned. This procedure has been adapted from Jiri's work but modified for the Coulomb gauge which was apparently wrong following some former implementation with RATIP.

JAC.InteractionStrength.MbaEmissionMigdalek — Method

InteractionStrength.MbaEmissionMigdalek(cp::CorePolarization, a::Orbital, b::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) interaction strength <b || O^(E1, emission with core-polarization) || a> in length gauge. A value::Float64 is returned.

JAC.InteractionStrength.XL_Breit — Method

InteractionStrength.XL_Breit(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid, eeint::Union{BreitInteraction, CoulombBreit, CoulombGaunt}; keep::Bool=false) ... computes the the effective Breit interaction strengths X^LBreit (abcd) or Gaunt interaction strengths X^LGaunt (abcd) for given rank L and orbital functions a, b, c and d at the given grid. For keep=true, the procedure looks up the (global) directory GBLStorageXL_Coulomb and returns the corresponding value without re-calculation of the interaction strength; it also 'stores' the calculated value if not yet included. For keep=false, the interaction strength is always computed on-fly. A value::Float64 is returned. At present, only the zero-frequency Breit or Gaunt interaction is taken into account.

JAC.InteractionStrength.XL_BreitDamped — Method

InteractionStrength.XL_BreitDamped(tau::Float64, L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid) ... computes the the effective Breit interaction strengths X^L_Breit (abcd) for given rank L and orbital functions a, b, c and d at the given grid. A value::Float64 is returned. At present, only the zero-frequency Breit interaction is taken into account.

JAC.InteractionStrength.XL_Breit_coefficients — Method

InteractionStrength.XL_Breit_coefficients(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital; onlyGaunt::Bool=false) ... evaluates the combinations and pre-coefficients for the zero-frequency Breit interaction X^L_Breit (omega=0.; abcd) for given rank L and orbital functions a, b, c and d. A list of coefficients xcList::Array{XLCoefficient,1} is returned.

JAC.InteractionStrength.XL_Breit_densities — Method

InteractionStrength.XL_Breit_densities(xcList::Array{XLCoefficient,1}, factor::Float64, grid::Radial.Grid) ... computes the the effective Breit interaction strengths X^L,0_Breit (abcd) for given rank L and a list of orbital functions a, b, c, d and angular coefficients at the given grid. A value::Float64 is returned. At present, only the zero-frequency Breit interaction is taken into account.

JAC.InteractionStrength.XL_Breit_reset_storage — Method

InteractionStrength.XL_Breit_reset_storage(keep::Bool; printout::Bool=false) ... resets the global storage of XL_Breit interaction strength; nothing is returned.

JAC.InteractionStrength.XL_Coulomb — Method

InteractionStrength.XL_Coulomb(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid; keep::Bool=false) ... computes the the effective Coulomb interaction strengths X^LCoulomb (abcd) for given rank L and orbital functions a, b, c and d at the given grid. For keep=true, the procedure looks up the (global) directory GBLStorageXLCoulomb and returns the corresponding value without re-calculation of the interaction strength; it also 'stores' the calculated value if not yet included. For keep=false, the interaction strength is always computed on-fly. A value::Float64 is returned.

JAC.InteractionStrength.XL_CoulombDamped — Method

InteractionStrength.XL_CoulombDamped(tau::Float64, L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid) ... computes the the effective Coulomb interaction strengths X^L_Coulomb (abcd) for given rank L and orbital functions a, b, c and d at the given grid. A value::Float64 is returned.

JAC.InteractionStrength.XL_Coulomb_DH — Method

InteractionStrength.XL_Coulomb_DH(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid, lambda::Float64) ... computes the the effective Coulomb-Debye-Hückel interaction strengths X^LCoulombDH (abcd) for given rank L and orbital functions a, b, c and d at the given grid and for the given screening parameter lambda. A value::Float64 is returned.

JAC.InteractionStrength.XL_Coulomb_WO — Method

InteractionStrength.XL_Coulomb_WO(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, grid::Radial.Grid) ... computes the the effective Coulomb interaction strengths X^L_Coulomb (abcd) for given rank L and orbital functions a, b, c and d at the given grid but without optimization. A value::Float64 is returned.

JAC.InteractionStrength.XL_Coulomb_reset_storage — Method

InteractionStrength.XL_Coulomb_reset_storage(keep::Bool; printout::Bool=false) ... resets the global storage of XL_Coulomb interaction strength; nothing is returned.

JAC.InteractionStrength.XL_plasma_ionSphere — Method

InteractionStrength.XL_plasma_ionSphere(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, lambda::Float64) ... computes the effective interaction strengths X^L_ion-sphere (abcd) for given rank L and orbital functions a, b, c and d and for the plasma parameter lambda. A value::Float64 is returned. Not yet implemented !

JAC.InteractionStrength.X_smsA — Method

InteractionStrength.X_smsA(a::Orbital, b::Orbital, c::Orbital, d::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the the effective interaction strengths X^1_sms,A (abcd) for fixed rank 1 and orbital functions a, b, c and d at the given grid. A value::Float64 is returned.

JAC.InteractionStrength.X_smsB — Method

InteractionStrength.X_smsB(a::Orbital, b::Orbital, c::Orbital, d::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the the effective interaction strengths X^1_sms,B (abcd) for fixed rank 1 and orbital functions a, b, c and d at the given grid. A value::Float64 is returned.

JAC.InteractionStrength.X_smsC — Method

InteractionStrength.X_smsC(a::Orbital, b::Orbital, c::Orbital, d::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the the effective interaction strengths X^1_sms,C (abcd) for fixed rank 1 and orbital functions a, b, c and d at the given grid. A value::Float64 is returned.

JAC.InteractionStrength.bosonShift — Method

InteractionStrength.bosonShift(a::Orbital, b::Orbital, potential::Array{Float64,1}, grid::Radial.Grid) ... computes the <a|| h^(boson-field) ||b> reduced matrix element of the boson-field shift Hamiltonian for orbital functions a, b. This boson-field shift Hamiltonian just refers to the effective potential of the given isotope due to the (assumed) boson mass. A value::Float64 is returned.

JAC.InteractionStrength.dipole — Method

InteractionStrength.dipole(a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| d ||b> reduced matrix element of the dipole operator for orbital functions a, b. A value::Float64 is returned.

JAC.InteractionStrength.eMultipole — Method

InteractionStrength.eMultipole(k::Int64, a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| t^(Ek) ||b> reduced matrix element of the dipole operator for orbital functions a, b. A value::Float64 is returned.

JAC.InteractionStrength.fieldShift — Method

InteractionStrength.fieldShift(a::Orbital, b::Orbital, deltaPotential::Array{Float64,1}, grid::Radial.Grid) ... computes the <a|| h^(field-shift) ||b> reduced matrix element of the field-shift Hamiltonian for orbital functions a, b. This field-shift Hamiltonian just refers to the difference of the nuclear potential deltaPotential of two isotopes, and which is already divided by the difference of the mean-square radii. A value::Float64 is returned.

JAC.InteractionStrength.hamiltonian_nms — Method

InteractionStrength.hamiltonian_nms(a::Orbital, b::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the <a|| h_nms ||b> reduced matrix element of the normal-mass-shift Hamiltonian for orbital functions a, b. A value::Float64 is returned. For details, see Naze et al., CPC 184 (2013) 2187, Eq. (37).

JAC.InteractionStrength.hfs_t1 — Method

InteractionStrength.hfs_t1(a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| t^(1) ||b> reduced matrix element for the HFS coupling to the magnetic-dipole moment of the nucleus for orbital functions a, b. A value::Float64 is returned.

JAC.InteractionStrength.hfs_t2 — Method

InteractionStrength.hfs_t2(a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| t^(2) ||b> reduced matrix element for the HFS coupling to the electric-quadrupole moment of the nucleus for orbital functions a, b. A value::Float64 is returned.

JAC.InteractionStrength.matrixL_Coulomb — Method

InteractionStrength.matrixL_Coulomb(L::Int64, a::Orbital, b::Orbital, c::Orbital, d::Orbital, primitives::Bsplines.Primitives) ... computes the partly-contracted (effective) Coulomb interaction matrices M^L_Coulomb (abcd) for given rank L and orbital functions a, b, c and d at the given grid. The matrix M^L is defined for the primitives and contracted over the two orbitals b, d (for a=c) or b, c (for a=d). An error message is issued if a != c && a != d. A matrix::Array{Float64,2} is returned.

JAC.InteractionStrength.multipoleTransition — Method

InteractionStrength.multipoleTransition(mp::EmMultipole, gauge::EmGauge, omega::Float64, b::Orbital, a::Orbital, grid::Radial.Grid) ... to compute the (single-electron reduced matrix element) multipole-transition interaction strength <b || T^(Mp, absorption) || a> due to Johnson (2007) for the interaction with the Mp multipole component of the radiation field and the transition frequency omega, and within the given gauge. A value::Float64 is returned.

JAC.InteractionStrength.schiffMoment — Method

InteractionStrength.schiffMoment(a::Orbital, b::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the <a|| h^(Schiff-moment) ||b> reduced matrix element of the Schiff-moment Hamiltonian for orbital functions a, b and for the nuclear density as given by the nuclear model. A value::Float64 is returned.

JAC.InteractionStrength.weakCharge — Method

InteractionStrength.weakCharge(a::Orbital, b::Orbital, nm::Nuclear.Model, grid::Radial.Grid) ... computes the <a|| h^(weak-charge) ||b> reduced matrix element of the weak-charge Hamiltonian for orbital functions a, b and for the nuclear density as given by the nuclear model. A value::Float64 is returned.

JAC.InteractionStrength.zeeman_Delta_n1 — Method

InteractionStrength.zeeman_Delta_n1(a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| Delta n^(1) ||b> reduced matrix element for the Zeeman-Schwinger contribution to the coupling to an external magnetic field for orbital functions a, b. A value::Float64 is returned.

JAC.InteractionStrength.zeeman_n1 — Method

InteractionStrength.zeeman_n1(a::Orbital, b::Orbital, grid::Radial.Grid) ... computes the <a|| n^(1) ||b> reduced matrix element for the Zeeman coupling to an external magnetic field for orbital functions a, b. A value::Float64 is returned.

Basics: data types

Basics: functions

Basics: compute & generate

B-Spline basis

JAC.BsplinesN.Bspline — Type

struct BsplinesN.Bspline ... defines a type for a (single) B-spline that is defined on a given radial grid from r[lower:upper]. Note that only the non-zero values are specified for the B-spline function and its derivative.

+ lower        ::Int64               ... lower radial index (on the radial grid.r) from where the functions is nonzero.
+ upper        ::Int64               ... upper radial index up to which the functions is nonzero.
+ bs           ::Array{Float64,1}    ... radial B-spline functions as defined on the predefined grid.r[lower:upper]
+ bp           ::Array{Float64,1}    ... derivative of bs on the predefined grid grid.r[lower:upper]

JAC.BsplinesN.Primitives — Type

struct BsplinesN.Primitives ... defines a type for a set of primitive functions which typically belongs to a well-defined grid.

+ grid         ::Radial.Grid         ... radial grid on which the states are represented.
+ bsplinesL    ::Array{Bspline,1}    ... set of B-splines for the large components on the given radial grid.
+ bsplinesS    ::Array{Bspline,1}    ... set of B-splines for the small components on the given radial grid.

JAC.BsplinesN.computeNondiagonalD — Method

BsplinesN.computeNondiagonalD(pm::Int64, kappa::Int64, bspline1::BsplinesN.Bspline, bspline2::BsplinesN.Bspline, grid::Radial.Grid) ... computes the (radial and non-diagonal) D_kappa^+/- integral two the bsplines, all defined on grid <bspline1| +/- d/dr + kappa/r | bspline2>. – pm = +1/-1 provides the phase for taking the derivative.

JAC.BsplinesN.computeOverlap — Method

BsplinesN.computeOverlap(bspline1::BsplinesN.Bspline, bspline2::BsplinesN.Bspline, grid::Radial.Grid) ... computes the (radial) overlap integral <bspline1|bsplines> for two bpslines as defined on grid.

JAC.BsplinesN.computeVlocal — Method

BsplinesN.computeVlocal(bspline1::BsplinesN.Bspline, bspline2::BsplinesN.Bspline, pot::Radial.Potential, grid::Radial.Grid) ... computes the (radial) integral <bspline1| V_pot |bsplines> for two bpslines and the given radial potential as defined on grid.

JAC.BsplinesN.extractBsplineCoefficients — Method

BsplinesN.extractBsplineCoefficients(sh::Subshell, wc::Basics.Eigen, grid::Radial.Grid) ... Here, it is assumed that the matrix wc contains the (column) eigenvectors as associated with a single-electron Dirac Hamiltonian matrix for symmetry kappa in Subshell(n, kappa). The procedure then extracts the (full) vector of B-spline coefficients for the radial orbital of subshell sh by applying the standard rules of atomic physics for the principal quantum number n. A vector::Array{Float64,1} is returned, whose length is nsL+nsS in the original basis of B-spline functions/primitives.

JAC.BsplinesN.generateGalerkinMatrix — Method

BsplinesN.generateGalerkinMatrix(sh::Subshell, energy::Float64, pot::Radial.Potential, primitives::BsplinesN.Primitives) ... generates the Galerkin-A matrix for the given potential and B-spline primitives; a matrix::Array{Float64,2} is returned.

JAC.BsplinesN.generateOrbitalFromPrimitives — Method

BsplinesN.generateOrbitalFromPrimitives(sh::Subshell, energy::Float64, mtp::Int64, ev::Array{Float64,1}, primitives::BsplinesN.Primitives) ... generates the large and small components of a (relativistic) orbital for the subshell sh from the given primitives and the eigenvector ev. A (non-normalized) orbital::Orbital is returned.

JAC.BsplinesN.generateOrbitalFromPrimitives — Method

BsplinesN.generateOrbitalFromPrimitives(sh::Subshell, wc::Basics.Eigen, primitives::BsplinesN.Primitives) ... generates the large and small components for the subshell sh from the primitives and their eigenvalues & eigenvectors. A (normalized) orbital::Orbital is returned.

JAC.BsplinesN.generateOrbitals — Method

BsplinesN.generateOrbitals(subshells::Array{Subshell,1}, pot::Radial.Potential, nm::Nuclear.Model, primitives::BsplinesN.Primitives; printout::Bool=true) ... generates all single-electron orbitals from subshell list for the radial potential pot. A set of orbitals::Dict{Subshell, Orbital} is returned.

JAC.BsplinesN.generateOrbitalsHydrogenic — Method

BsplinesN.generateOrbitalsHydrogenic(subshells::Array{Subshell,1}, nm::Nuclear.Model, primitives::BsplinesN.Primitives; printout::Bool=true) ... generates all single-electron orbitals from subshell list for the nuclear potential as specified by nm. A set of orbitals::Dict{Subshell, Orbital} is returned.

JAC.BsplinesN.generatePrimitives — Method

BsplinesN.generatePrimitives(grid::Radial.Grid) ... generates the breaks, knots and the B-spline primitives of order k, both for the large and small components. The function applies the given grid parameters; no primitive is defined beyond grid[n_max]. The definition of the primitives follow the work of Zatsarinny and Froese Fischer, CPC 202 (2016) 287. –- A (set of) primitives::BsplinesN.Primitives is returned.

JAC.BsplinesN.generateTTpMatrix! — Method

BsplinesN.generateTTpMatrix!(TTp::String, kappa::Int64, primitives::BsplinesN.Primitives, storage::Dict{String,Array{Float64,2}}) ... returns the TTp block of the (single-electron) Dirac Hamiltonian matrix for an electron with symmetry kappa without any potential. The following TTp strings are allowed: ["LL-overlap", "SS-overlap", "LS-Dkappa^-", "LS-Dkappa^+"].

    Two modes are distinguished owing to the values that are available in the storage (Dict).
        * The TTp matrix block from the storage is returned, if an entry is known; it is assumed that this matrix
          block belong to the given set of primitives.
        * The TTp matrix is computed and set to the storage otherwise; from the TTp string, the key string
          key = string(kappa) * ":" * TTp is generated an applied in the storage dictionary.
          
    All B-splines are supposed to be defined for the same (radial) grid; a  matrix::Array{Float64,2}  is returned which 
    is quadratic for 'LL-overlap' and 'SS-overlap' and whose dimension depends on the number of B-splines for the large 
    and small component, otherwise.

JAC.BsplinesN.setupLocalMatrix — Method

`BsplinesN.setupLocalMatrix(kappa::Int64, primitives::BsplinesN.Primitives, pot::Radial.Potential, storage::Dict{String,Array{Float64,2}}) ...set-up the local parts of the generalized eigenvalue problem for the symmetry block kappa and the given (local) potential pot. The B-spline (basis) functions are defined by primitivesL for the large component and primitivesS for the small one, respectively.

Electron continuum

JAC.Continuum.Settings — Type

struct Continuum.Settings ... defines a type for the parameters for computing continuum orbitals.

+ includeExchange         ::Bool    ... True, if the exchange is to be included and false otherwise.
+ mtp                     ::Int64   ... No of grid points for which the continuum orbital(s) are to be computed.

JAC.Continuum.generateOrbitalAsymptoticCoulomb — Method

Continuum.generateOrbitalAsymptoticCoulomb(energy::Float64, sh::Subshell, pot::Radial.Potential, settings::Continuum.Settings) ... to generate a simple relativistic and non-normalized continuum orbital but by assuming an asymptotic sin/cos behaviour for both components at all mesh points. For this behaviour, both components and their derivatives can be calculated analytically. The effective charge Zbar is determined for the given potential and the non-Coulombic phase shifts is set to zero. An non-normalized cOrbital::Orbital is returned.

    Warning: At present, only the real part eta.re is taken into account; check for consistency !!

JAC.Continuum.generateOrbitalBessel — Method

Continuum.generateOrbitalBessel(energy::Float64, sh::Subshell, grid::Radial.Grid, settings::Continuum.Settings) ... to generate a simple, non-relativistic and non-normalized Bessel wave for the large component of the continuum orbital and a small component due to the kinetic-balance condition. A (non-normalized) orbital::Orbital is returned. While Pprime is obtained from the analytical expression of the large component, Qprime is set to zero here.

JAC.Continuum.generateOrbitalForLevel — Method

Continuum.generateOrbitalForLevel(energy::Float64, sh::Subshell, level::Level, nm::Nuclear.Model, grid::Radial.Grid, basis::Basis, settings::Continuum.Settings) ... to generate a continuum orbital for the (continuum) subshell sh, the energy and the effective charge within the given potential. The continuum orbital is generated orthogonal with regard to all subshells of the same symmetry in the basis. All further specifications about this generations are made by proper settings. A tupel of a (continuum) (orbital::Orbital, phase::Float64) is returned.

JAC.Continuum.generateOrbitalGalerkin — Method

Continuum.generateOrbitalGalerkin(energy::Float64, sh::Subshell, pot::Radial.Potential, settings::Continuum.Settings) ... to generate a non-normalized continuum orbital within the given local potential by using the Galerkin method and a given B-spline basis. A (non-normalized) orbital::Orbital is returned.

JAC.Continuum.generateOrbitalLocalPotential — Method

Continuum.generateOrbitalLocalPotential(energy::Float64, sh::Subshell, pot::Radial.Potential, settings::Continuum.Settings) ... to generate a continuum orbital for the (continuum) subshell sh, the energy and the given potential. The effective charge for the normalization of the continuum orbital is derived from the potential. All further specifications about this generations are made by proper settings; however, the function termintates if the settings.includeExchange = true. A tupel of a (continuum) (orbital::Orbital, phase::Float64, normFactor::Float64) is returned.

JAC.Continuum.generateOrbitalNonrelativisticCoulomb — Method

Continuum.generateOrbitalNonrelativisticCoulomb(energy::Float64, sh::Subshell, Zeff::Float64, grid::Radial.Grid, settings::Continuum.Settings) ... to generate a simple, non-relativistic and non-normalized Coulomb wave for the (continuum) subshell sh, the small component is simply obtained by the kinetic-balance condition. A orbital::Orbital is returned. ***** This function does not yet work because there is no Julia implementation of the hypergeometric function with complex arguments available.

JAC.Continuum.generateOrbitalNonrelativisticCoulomb — Method

(energy::Float64, sh::Subshell, pot::Radial.Potential, settings::Continuum.Settings) ... to generate a simple, non-relativistic and non-normalized Coulomb wave for the (continuum) subshell sh, the small component is simply obtained by the kinetic-balance condition. A orbital::Orbital is returned.

JAC.Continuum.generateOrbitalPureSine — Method

Continuum.generateOrbitalPureSine(energy::Float64, sh::Subshell, grid::Radial.Grid, settings::Continuum.Settings) ... to generate a simple, non-relativistic and non-normalized Bessel wave for the large component of the continuum orbital and a small component due to the kinetic-balance condition. A (non-normalized) orbital::Orbital is returned. While Pprime is obtained from the analytical expression of the large component, Qprime is set to zero here.

JAC.Continuum.gridConsistency — Method

Continuum.gridConsistency(maxEnergy::Float64, grid::Radial.Grid) ... to check the consistency of the given grid with the maximum energy of the required continuum electrons; an error message is issued if the grid.hp = 0. or 15 * grid.hp < wavelength(maxEnergy) or if the grid has less than 600 grid points. The function also returns the recommended grid point where the normalization and phase is to be determined. This number is currently set to nrContinuum = grid.NoPoints - 200 ... to correct for the wrong 'phase behaviour' at large r-values.

JAC.Continuum.normalizeOrbitalAlok — Method

Continuum.normalizeOrbitalAlok(cOrbital::Orbital, pot::Radial.Potential, settings::Continuum.Settings) ... to normalize the given continuum orbital with regard to a (asymptotic) wave function as per Salvat Code. The orbitasl are normalized to unit amplitude. An ( orbital::Orbital, (δ + Δ)::Float64, N::Float64 ) is returned.

JAC.Continuum.normalizeOrbitalCoulombSine — Method

Continuum.normalizeOrbitalCoulombSine(cOrbital::Orbital, pot::Radial.Potential, settings::Continuum.Settings) ... to normalize the given continuum orbital with regard to a (asymptotic) pure sine-function. An (on-energy-scale-normalized) orbital::Orbital is returned.

JAC.Continuum.normalizeOrbitalOngRussek — Method

Continuum.normalizeOrbitalOngRussek(cOrbital::Orbital, pot::Radial.Potential, settings::Continuum.Settings) ... to normalize the given continuum orbital with regard to a (asymptotic) pure sine-function. An (on-energy-scale-normalized) orbital::Orbital is returned.

JAC.Continuum.normalizeOrbitalPureSine — Method

Continuum.normalizeOrbitalPureSine(cOrbital::Orbital, grid::Radial.Grid, settings::Continuum.Settings) ... to normalize the given continuum orbital with regard to a (asymptotic) pure sine-function. An (on-energy-scale-normalized) orbital::Orbital and its relative phase phi w.r.t. sin(kr + phi) is returned.

JAC.Continuum.twoFzero — Method

function twoFzero(CA::ComplexF64, CB::ComplexF64, CZ::ComplexF64) ... Calculates the Hypergeometric function 2F0(CA,CB;1/CZ) hypergeometric asymptotic series. Taken from Radial package by Salvat et al. A ComplexF64 value is returned.

Hydrogenic ions

JAC.HydrogenicIon.energy — Method

HydrogenicIon.energy(sh::Shell, Z::Float64) ... to compute the (non-relativistic) energy for the given shell and for a point-like nuclear charge Z; the energy is printed in the current energy units to screen but is returned in Hartree. A energy::Float64 is returned.

JAC.HydrogenicIon.energy — Method

HydrogenicIon.energy(sh::Subshell, Z::Float64) ... to computes the Dirac energy for the hydrogenic subshell sh and for point-like nucleus with nuclear charge Z; a energy::Float64 in atomic units and without the rest energy of the electron is returned. That is the binding energy of a 1s_1/2 electron for Z=1 is -0.50000665.

JAC.HydrogenicIon.orbital — Method

HydrogenicIon.orbital(sh::Subshell, nm::Nuclear.Model, grid::Radial.Grid) ... to compute a relativstic hydrogenic Dirac orbital for the given nuclear model by using an explicit diagonalization of the Dirac Hamiltonian in a B-spline basis; an orbital::Radial.Orbital is returned.

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Shell, Z::Float64, r::Float64) ... to compute the (non-relativistic) orbital function P(r) for the given shell and nuclear charge Z; a value::Float64 is returned.

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Shell, Z::Float64, grid::Radial.Grid) ... to compute the same but for all r-values as specified by the given grid; a PList::Array{Float64,1} is returned.

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Shell, Z::Float64, rlist::Array{Float64,1}) ... to compute the same but for an array of r-values [r1, r2, ...]; a PList::Array{Float64,1} is returned.

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Subshell, Z::Float64, r::Float64) ... to compute a relativistic hydrogenic Dirac orbital for the given subshell and nuclear charge Z; a value::Float64 is returned; contributed by C Naumann (2022). Implementation of the analytical solution of the Dirac equation for a point-like nucleus following the book by Johnson: Atomic Structure Theory.

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Subshell, Z::Float64, grid::Radial.Grid) ... to compute a relativistic hydrogenic Dirac orbital on the given grid; an Array with the large and small component is returned; contributed by C Naumann (2022).

JAC.HydrogenicIon.radialOrbital — Method

HydrogenicIon.radialOrbital(sh::Subshell, nm::Nuclear.Model, grid::Radial.Grid) ... to compute a relativstic hydrogenic Dirac orbital for the given nuclear model by using an explicit diagonalization of the Dirac Hamiltonian in a B-spline basis; an orbital::Radial.Orbital is returned; contributed by C Naumann (2022).

JAC.HydrogenicIon.radialOrbital_old2022 — Method

HydrogenicIon.radialOrbital_old2022(sh::Subshell, Z::Float64, grid::Radial.Grid) ... to compute a relativstic hydrogenic Dirac orbital on the given grid by applying the kinetic-balance to a corresponding non-relavistic orbital; an orbital::Radial.Orbital is returned.

JAC.HydrogenicIon.radialOrbital_old2022 — Method

HydrogenicIon.radialOrbital_old2022(sh::Subshell, nm::Nuclear.Model, grid::Radial.Grid) ... to compute a relativstic hydrogenic Dirac orbital for the given nuclear model by using an explicit diagonalization of the Dirac Hamiltonian in a B-spline basis; an orbital::Radial.Orbital is returned.

JAC.HydrogenicIon.rkExpectation — Method

HydrogenicIon.rkExpectation(srk::String, sh::Shell, Z::Float64) ... to compute the (non-relativistic) r^k expectation value for the shell sh of an ion with charge Z; a value::Float64 [in a_o^k] is returned. The string can takes values srk = ["r", "r^2", "1/r"]

Spin angular coefficients

JAC.SpinAngular.AbstractAngularType — Type

abstract type SpinAngular.AbstractAngularType ... defines an abstract type and a number of data types to work with one- and two-particle operators of given rank, see also:

+ struct OneParticleOperator    ... to represent a single-particle operator with well-defined spherical tensor rank.
+ struct TwoParticleOperator    ... to represent a two-particle operator with well-defined spherical tensor rank.

JAC.SpinAngular.Coefficient1p — Type

struct SpinAngular.Coefficient1p ... a struct for defining a single spin-angular coefficient for a reduced one-particle matrix element <a || o^(L) || b>.

+ nu       ::Int64             ... Rank (L or nu) of the single-particle interaction strength.
+ a        ::Subshell          ... Left-hand subshell (orbital).
+ b        ::Subshell          ... Right-hand subshell (orbital).
+ T        ::Float64           ... (Value of) spin-angular coefficient.

JAC.SpinAngular.Coefficient2p — Type

struct SpinAngular.Coefficient2p ... a struct for defining a single spin-angular coefficient for a reduced two-particle matrix element <ab || o^(L) || cd>, such as the Coulomb interaction strength.

+ nu       ::Int64             ... Rank (L or nu) of the single-particle interaction strength.
+ a        ::Subshell          ... Left-hand subshell (orbital).
+ b        ::Subshell          ... Left-hand subshell (orbital).
+ c        ::Subshell          ... Right-hand subshell (orbital).
+ d        ::Subshell          ... Right-hand subshell (orbital).
+ V        ::Float64           ... (Value of) spin-angular coefficient.

JAC.SpinAngular.DiagramC01 — Type

struct SpinAngular.Diagram ... to defines various singleton() structs in order to distinguish between different coupling schemes of the matrix elements.

JAC.SpinAngular.OneParticleOperator — Type

struct SpinAngular.OneParticleOperator <: AbstractAngularType ... a struct for defining the spherial tensor structure of a single-particle operator.

+ rank            ::Int64             ... Rank of the operator.
+ parity          ::Basics.Parity     ... Parity of the operator (if needed ??)
+ sameOrbitalSet  ::Bool              ... True if orbitals for both CSF are taken from the same orbital set (if needed ??)

JAC.SpinAngular.OneParticleOperator — Method

SpinAngular.OneParticleOperator() ... constructor for setting the default values.

JAC.SpinAngular.QspaceTerm — Type

struct SpinAngular.QspaceTerm ... a struct for defining a subshell term/state |j (nu) alpha Q J> == |j (nu) Q J Nr> for a subshell with well-defined j.

+ j        ::AngularJ64   ... subshell j
+ Q        ::AngularJ64   ... quasi-spin
+ J        ::AngularJ64   ... total J of subshell term
+ Nr       ::Int64        ... Additional quantum number Nr = 0,1,2.
+ min_odd  ::Int64        ... the minimal limits of the subshell terms for odd number operators in second quantization
+ max_odd  ::Int64        ... the maximal limits of the subshell terms for odd number operators in second quantization
+ min_even ::Int64        ... the minimal limits of the subshell terms for even number operators in second quantization
+ max_even ::Int64        ... the maximal limits of the subshell terms for even number operators in second quantization

JAC.SpinAngular.SchemeEta_a — Type

struct SpinAngular.SchemeEta ... to define various singleton structs in order to distinguish between different irreducible tensors and matrix elements.

+ Eta_a         ... a^(qj)_mq
+ Eta_W         ... W^(k_12) = [a x a]
+ Eta_aW        ... W^(k_12, k_2) = [a1 x [a2 x a3]]
+ Eta_Wa        ... W^(k_12, k_2) = [[a1 x a2] x a3]
+ Eta_WW        ... W^(kk0) = [[a1 x a2]^k x [a3 x a4]^k]^0

JAC.SpinAngular.TwoParticleOperator — Type

struct SpinAngular.TwoParticleOperator ... a struct for defining the spherial tensor structure of a two-particle operator.

+ rank            ::Int64             ... Rank of the operator (if needed ??).
+ parity          ::Basics.Parity     ... Parity of the operator (if needed ??)
+ sameOrbitalSet  ::Bool              ... True if orbitals for both CSF are taken from the same orbital set (if needed ??)

JAC.SpinAngular.TwoParticleOperator — Method

SpinAngular.TwoParticleOperator() ... constructor for setting the default values.