Defaults

JAC.Defaults.convertUnits — Method

Defaults.convertUnits() ... converts some data from one format/unit into another one; cf. the supported keystrings and return values.

("cross section: from atomic to predefined unit", value::Float64) or ("cross section: from atomic", value::Float64) ... to convert an cross section value from atomic to the predefined cross section unit; a Float64 is returned.

("cross section: from barn to atomic unit", value::Float64) ... to convert an cross section value from barn atomic section unit; a Float64 is returned.

("cross section: from atomic to barn", value::Float64) or ("cross section: from atomic to Mbarn", value::Float64) or ("cross section: from atomic to cm^2", value::Float64) ... to convert an energy value from atomic to the speficied cross section unit; a Float64 is returned.

("cross section: from predefined to atomic unit", value::Float64) or ("cross section: to atomic", value::Float64) ... to convert a cross section value from the predefined to the atomic cross section unit; a Float64 is returned.

("einstein B: from atomic", value::Float64) ... to convert a Einstein B coefficient from atomic to the speficied energy units; a Float64 is returned.

("density: from [g/cm^3] to atomic", value::Float64) ... to convert a mass density from [g/cm^3] to atomic units [u/a_o^3]; a Float64 is returned.

("energy-diff. cross section: from atomic to predefined unit", value::Float64) or ("energy-diff. cross section: from atomic", value::Float64) ... to convert an energy-diff. cross section value from atomic to the predefined energy-diff. cross section unit; a Float64 is returned.

("energy: from atomic to eV", value::Float64) or ("energy: from atomic to Kayser", value::Float64) or ("energy: from atomic to Hz", value::Float64) or ("energy: from atomic to Angstrom", value::Float64) or ("energy: from atomic to Ws", value::Float64) ... to convert an energy value from atomic to the speficied energy unit; a Float64 is returned.

("energy: from predefined to atomic unit", value::Float64) or ("energy: to atomic", value::Float64)... to convert an energy value from the predefined to the atomic energy unit; a Float64 is returned.

("energy: from eV to atomic", value::Float64) ... to convert an energy value from eV to the atomic energy unit; a Float64 is returned.

("energy: from wavelength [nm] to atomic", value::Float64) ... to convert a wavelength [nm] to the atomic energy unit; a Float64 is returned.

("intensity: from W/cm^2 to atomic", value::Float64) ... to convert the intensity [in W/cm^2] to the atomic intensity unit; a Float64 is returned.

("kinetic energy to wave number: atomic units", value::Float64) ... to convert a kinetic energy value (in a.u.) into a wave number k (a.u.); a Float64 is returned.

("kinetic energy to wavelength: atomic units", value::Float64) ... to convert a kinetic energy value (in a.u.) into a wavelength (a.u.); a Float64 is returned.

("length: from fm to atomic", value::Float64) ... to convert a length value (in fm) into a.u.; a Float64 is returned.

("length: from atomic to fm", value::Float64) or ("energy: from atomic to Kayser", value::Float64) ... to convert a length value (in Bohr's a.u.) to the speficied length unit; a Float64 is returned.

("rate: from atomic to predefined unit", value::Float64) or ("rate: from atomic", value::Float64) ... to convert a rate value from atomic to the predefined rate unit; a Float64 is returned.

("rate: from atomic to 1/s", value::Float64) ... to convert an rate value from atomic to the speficied rate unit; a Float64 is returned.

("rate: from predefined to atomic unit", value::Float64) or `("rate: to atomic", value::Float64)'... to convert a rate value from the predefined to the atomic rate unit; a Float64 is returned.

("strength: from atomic to predefined unit", value::Float64) or ("strength: from atomic", value::Float64) ... to convert a (resonance) strength value from atomic to the predefined rate unit; a Float64 is returned.

("time: from atomic to predefined unit", value::Float64) or ("time: from atomic", value::Float64) ... to convert an time value from atomic to the predefined time unit; a Float64 is returned.

("time: from atomic to sec", value::Float64) or ("time: from atomic to fs", value::Float64)' or("time: from atomic to as", value::Float64)` ... to convert a time value from atomic to the speficied time unit; a Float64 is returned.

("time: from predefined to atomic unit", value::Float64) or `("time: to atomic", value::Float64)' ... to convert a time value from the predefined to the atomic time unit; a Float64 is returned.

("temperature: from Kelvin to (Hartree) units", value::Float64) ... to convert a temperature in Kelvin into atomic (Hartree) units; a Float64 is returned.

("temperature: from atomic to Kelvin", value::Float64) ... to convert an atomic (energy) unit into Kelvin; 1 Hartree = 315774.64 K; a Float64 is returned.

("wave number to total electron energy: atomic units", value::Float64) ... to convert a wavenumber (a.u.) into the total electron energy, including the rest energy; a Float64 is returned.

("wave number to kinetic energy: atomic units", value::Float64) ... to convert a wavenumber (a.u.) into the kinetic energy; a Float64 is returned.

JAC.Defaults.convertUnits — Method

(string, values::Array{Float64,1}) ... to convert for the same strings as above but for an list of values; a corresponding Array{Float64,1} is returned.

JAC.Defaults.getDefaults — Method

("ordered shell list: non-relativistic", n_max::Int64) ... to give an ordered list of non-relativistic shells::Array{Shell,1} up to the (maximum) principal number n_max.

("ordered subshell list: relativistic", n_max::Int64) ... to give an ordered list of relativistic subshells::Array{Subshell,1} up to the (maximum) principal number n_max.

JAC.Defaults.getDefaults — Method

Defaults.getDefaults() ... gives/supplies different information about the (present) framework of the computation or about some given data; cf. Defaults.setDefaults().

("alpha") or ("fine-structure constant alpha") ... to get the (current) value::Float64 of the fine-structure constant alpha.

("electron mass: kg") or ("electron mass: amu") ... to get the (current) value::Float64 of the electron mass in the specified unit.

("framework") ... to give the (current) setting::String of the overall framework.

("electron rest energy") or ("mc^2") ... to get the electron rest energy.

("electron g-factor") ... to give the electron g-factor g_s = 2.00232.

("unit: energy") or ("unit: cross section") or ("unit: rate") or ("unit: strength") or ("unit: time") ... to get the corresponding (user-defined) unit::String for the current computations.

("standard grid") ... to get the (current standard) grid::Array{Float64,1} to which all radial orbital functions usually refer.

("speed of light: c") ... to get the speed of light in atomic units.

("summary flag/stream") ... to get the logical flag and stream for printing a summary file; a tupel (flag, iostream) is returned.

JAC.Defaults.saRatip — Method

Defaults.saRatip() ... returns true or false if the spin-angular coefficients should be calculated by Fortran codes from RATIP/Grasp.

JAC.Defaults.setDefaults — Method

("QED: damped-hydrogenic", Znuc::Float64, wa::Array{Float64,1}) ... to (re-) define the lambda-C damped overlap integrals of the lowest kappa-orbitals [ wa1s1/2, wa2p1/2, wa2p3/2, wa3d3/2, wa3d5/2 ] for the (new) nuclear charge Znuc; nothing is returned.

JAC.Defaults.setDefaults — Method

("continuum: potential", scField::Basics.AbstractScField) ... to (re-) define the potential that is applied for the generation of the continuum orbitals.

JAC.Defaults.setDefaults — Method

("standard grid", grid::Radial.Grid; printout::Bool=true) ... to (pre-) define internally the standard radial grid which is used to represent most orbitals.

JAC.Defaults.setDefaults — Method

("relativistic subshell list", subshells::Array{Subshell,1}; printout::Bool=true) ... to (pre-) define internally the standard relativistic subshell list on which the standard order of orbitals is based.

JAC.Defaults.setDefaults — Method

Defaults.setDefaults() ... (re-) defines some 'standard' settings which are common to all the computations with the JAC module, and which can be 'overwritten' by the user. –- An improper setting of some variable may lead to an error message, if recognized immediately. The following defaults apply if not specified otherwise by the user: the framework is 'relativistic', energies are given in eV and cross sections in barn. Note that, internally, atomic units are used throughout for all the computations within the program. nothing is returned if not indicated otherwise.

+ ("framework: relativistic") or ("framework: non-relativistic")

... to define a relativistic or non-relativistic framework for all subsequent computations.

("method: continuum, spherical Bessel") or ("method: continuum, pure sine") or ("method: continuum, asymptotic Coulomb") or ("method: continuum, nonrelativistic Coulomb") or ("method: continuum, Galerkin") ... to define a a method for the generation of the continuum orbitals as (pure) spherical Bessel, pure sine, asymptotic Coulomb, nonrelativistic Coulomb orbital or by means of the B-spline-Galerkin method.

("method: normalization, pure sine") or ("method: normalization, pure Coulomb") or ("method: normalization, Ong-Russek") ... to define a method for the normalization of the continuum orbitals as asymptotically (pure) sine or Coulomb functions, or following the procedure by Ong & Russek (1978).

("QED model: Petersburg") or ("QED model: Sydney") ... to define a model for the computation of the QED corrections following the work by Shabaev et al. (2011; Petersburg) or Flambaum and Ginges (2004; Syney).

("unit: energy", "eV") or ("unit: energy", "Kayser") or ("unit: energy", "Hartree") or ("unit: energy", "Hz") or ("unit: energy", "Hz") ... to (pre-) define the energy units for all further printouts and communications with the JAC module.

("unit: cross section", "a.u.") or ("unit: cross section", "barn") or ("unit: cross section", "Mbarn") ... to (pre-) define the unit for the printout of cross sections.

("unit: rate", "a.u.") or ("unit: rate", "1/s") ... to (pre-) define the unit for the printout of rates.

("unit: resonance strength", "a.u.") or ("unit: resonance strength", "barn eV") or ("unit: resonance strength", "cm^2 eV") ... to (pre-) define the unit for the printout of resonance strengths.

("unit: time", "a.u.") or ("unit: time", "sec") or ("unit: time", "fs") or ("unit: time", "as") ... to (pre-) define the unit for the printout and communications of times with the JAC module.

JAC.Defaults.warn — Method

Defaults.warn() ... deals with warnings that occur during a run and session; it handles the global array GBL_WARNINGS.

(AddWarning, warning::String) ... to add warning to the global array GBL_WARNINGS.

(PrintWarnings) ... to print all warnings that are currently kept in the global array GBL_WARNINGS.

(ResetWarnings) ... to reset the global array GBL_WARNINGS.

JAC.Radial

JAC.Radial.OrbitalBunge1993 — Method

Radial.OrbitalBunge1993(subshell::Subshell, Z::Int64) ... to calculate the radial orbital on the standard grid.

JAC.Radial.OrbitalPrimitiveSlater — Method

Radial.OrbitalPrimitiveSlater(subshell::Subshell, N::Int64, alpha0::Float64, beta0::Float64, grid::Radial.Grid) ... to calculate a list of (non-relativistic) radial Slater primitives::Array{Orbital,1} for the subshell-symmetry on the given grid. All orbitals have the same subshell, isBound = true, useStandardGrid = true and energy = 0. The small components are constructed by applying kinetic balance to the large component, and all the primitives are properly normalized.

JAC.Radial.checkGrid — Method

Radial.checkGrid(grid::Radial.Grid, shell::Shell, Z::Float64) ... checks that the radial grid can represent the hydrogenic orbitals for the given subshell of nuclear charge Z; it simply tests that the boxSize > 4 * n^2 / Z [a_o]. (More accurate one could use the mean expectation values of r of this shell). The program reports about this comparison and terminates if this is not the case. Nothing is returned.

JAC.Radial.checkGrid — Method

Radial.checkGrid(grid::Radial.Grid, subshell::Subshell, relAccuracy::Float64) ... checks that the radial grid can represent the hydrogenic orbitals for the given subshell with a relative accuracy <= relAccuracy. The program reports about the relative accuracy and terminates with an error message if this is not the case. Nothing is returned.

JAC.Radial.compute_McLean1981 — Method

Radial.OrbitalMcLean1981(subshell::Subshell,Z::Int64) ... to calculate the radial orbital on the standard grid.

JAC.Radial.determineGrid — Method

Radial.determineGrid(grid::Radial.Grid; printout::Bool=false) ... determines the detailed radial grid due to the given gridType and parameters; a gr::Radial.Grid is returned

JAC.Radial.determineNoPoints — Method

Radial.determineNoPoints(rnt::Float64, h::Float64, hp::Float64, rbox::Float64, orderGL::Int64) ... determines the number of points of the physical size if the grid parameters and the box-size is given; moreover, it is ensured that this NoPoints is consistent with the largest break point for the B-spline grid. An NoPoints::Int64 is returned.

JAC.Radial.determineZbar — Method

Radial.determineZbar(pot::Radial.Potential) ... determines the effective charge that is asymptotically seen by the electron in the potential pot. A Zbar::Float64 is returned.

JAC.Radial.generateGrid — Method

`Radial.generateGrid(grid::Radial.Grid; boxSize::Union{Nothing,Float64}=nothing, maximumFreeElectronEnergy::Union{Nothing,Float64}=nothing, maximumPrincipalQN::Union{Nothing,Int64}=nothing, NoPointsInsideNucleus::Union{Nothing,Int64}=nothing, NoPointsInsideFirstBohrRadius::Union{Nothing,Int64}=nothing)

... to generate a grid that fulfills special requirements; th following schemes are supported:

    boxSize 
        ... apply a fixed box size (in atomic units) as needed in an average-atom model and elsewhere.
    boxSizeWithZeroWeights 
        ... apply a fixed box size (in atomic units) as needed in an average-atom model and elsewhere.
            Here, a standard grid is used but with all weights w [r] = 0 for r > boxSize.
    maximumFreeElectronEnergy
        ... provide a maximum free-electron energy (Hartree) and determine the linearized stepsize such,
            that 20 points per wavelength are used asymptotically.
    maximumPrincipalQN
        ... generate an grid with a rbox-size that is suitable to represent subshell orbitals
            with the given n; the value of rbox = 5 * <r_n> is taken, i.e. 5 times the mean hydrogenic 
            value (not yet).
    NoPointsInsideNucleus
        ... generate a grid (of given type) with the given number inside the nucleus; this affects the values
            of rnt and h (not yet).
    NoPointsInsideFirstBohrRadius
        ... generate a grid (of given type) with the given number inside the first Bohr radius; this affects 
            the values of rnt and h (not yet).

Only on of these optional parameters can be selected at a given time. A proper grid::Radial.Grid is returned,
along with a short reasoning of what has been selected.

JAC.ImpactExcitation

JAC.ImpactExcitation.amplitude — Method

ImpactExcitation.amplitude(kind::AbstractEeInteraction, channel::ImpactExcitation.Channel, cFinalLevel::Level, cInitialLevel::Level, grid::Radial.Grid; printout::Bool=true) ... to compute the kind in CoulombInteraction(), BreitInteraction(), CoulombBreit() electron-impact interaction amplitude <(alphaf Jf, kappaf) Jt || O^(e-e, kind) || (alphai Ji, kappai) Jt> due to the interelectronic interaction for the given final and initial (continuum) level. A value::ComplexF64 is returned.

JAC.ImpactExcitation.computeAmplitudesProperties — Method

ImpactExcitation.computeAmplitudesProperties(line::ImpactExcitation.Line, nm::Nuclear.Model, grid::Radial.Grid, settings::ImpactExcitation.Settings; printout::Bool=true) ... to compute all amplitudes and properties of the given line; a line::ImpactExcitation.Line is returned for which the amplitudes and properties are now evaluated.

JAC.ImpactExcitation.computeEffStrengths — Method

ImpactExcitation.computeEffStrengths(lines::Array{ImpactExcitation.Line, 1}, settings::ImpactExcitation.Settings) ... computes Effective collision strengths from the calculated line collision strengths at temperature(s) [K]. Returns an Array{ImpactExcitation.RateCoefficients,1}.

JAC.ImpactExcitation.computeLines — Method

ImpactExcitation.computeLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::ImpactExcitation.Settings; output=true) ... to compute the electron-impact excitation transition amplitudes and all properties as requested by the given settings. A list of lines::Array{ImpactExcitation.Lines} is returned.

JAC.ImpactExcitation.computeRateCoefficients — Method

ImpactExcitation.computeRateCoefficients(effStrengths::Vector{RateCoefficients}, settings::ImpactExcitation.Settings) ... computes Exitation rate coefficients from the calculated collsion strengths at a temperature(s) [K]. The rate coefficients are returned in [cm^3/s]. Returns an Array{ImpactExcitation.RateCoefficients,1}.

JAC.ImpactExcitation.determineChannels — Method

ImpactExcitation.determineChannels(finalLevel::Level, initialLevel::Level, settings::ImpactExcitation.Settings) ... to determine a list of electron-impact excitation Channels for a transitions from the initial to the final level and by taking into account the particular settings of for this computation; an Array{ImpactExcitation.Channel,1} is returned.

JAC.ImpactExcitation.determineLines — Method

ImpactExcitation.determineLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, settings::ImpactExcitation.Settings) ... to determine a list of ImpactExcitation.Line's for transitions between levels from the initial- and final-state multiplets, and by taking into account the particular selections and settings for this computation; an Array{ImpactExcitation.Line,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JAC.ImpactExcitation.displayLines — Method

ImpactExcitation.displayLines(lines::Array{ImpactExcitation.Line,1}) ... to display a list of lines and channels that have been selected due to the prior settings. A neat table of all selected transitions and energies is printed but nothing is returned otherwise.

JAC.ImpactExcitation.displayResults — Method

ImpactExcitation.displayResults(line::ImpactExcitation.Line) ... to display energies, cross sections, collision stregths, convergence etc. for a single line immediatly after the calculation for that line is completed.

JAC.ImpactExcitation.displayResults — Method

ImpactExcitation.displayResults(lines::Array{ImpactExcitation.Line,1}) ... to list all results, energies, cross sections, etc. of the selected lines. A neat table is printed but nothing is returned otherwise.

JAC.ImpactExcitation.displayResults — Method

ImpactExcitation.displayResults(allRates::Array{RateCoefficients,1}) ... to list the excitation rate coefficients and effective collision strengths for the selected lines at the selected temperatures. A neat table is printed but nothing is returned otherwise.

JAC.ImpactExcitation.groupLines — Method

ImpactExcitation.groupLines(lines::Array{ImpactExcitation.Line,1}, settings::ImpactExcitation.Settings) ... groups lines having the same initial and final level but different energies returns an Array{Array{ImpactExcitation.Line,1},1}

JAC.ImpactExcitation.groupLines — Method

ImpactExcitation.groupLines(lines::Array{ImpactExcitation.Line,1}) ... groups lines having the same initial and final level but different energies returns an Array{Array{ImpactExcitation.Line,1},1}

JAC.ImpactExcitation.interpolateCS — Method

ImpactExcitation.interpolateCS(x::Float64, xa::Vector{Float64}, ya::Vector{Float64}, isE1Allowed::Bool) ... interpolates the cross section or collision stregths for a given initial electron Energy if the energy is beyond the upper bound then uses asymptotic approximation for extrapolation, depending on the transition is E1 allowed or not. Returns a Float64.

JAC.ImpactExcitation.interpolateCS — Method

ImpactExcitation.interpolateCS(x::Float64, xa::Vector{Float64}, ya::Vector{Float64}) ... interpolates the cross section or collision stregths for a given initial electron Energy returns a Float64.

JAC.PhotoExcitation

JAC.PhotoExcitation.computeAmplitudesProperties — Method

PhotoExcitation.computeAmplitudesProperties(line::PhotoExcitation.Line, grid::Radial.Grid, settings::PhotoExcitation.Settings; printout::Bool=true) ... to compute all amplitudes and properties of the given line; a line::PhotoExcitation.Line is returned for which the amplitudes and properties have now been evaluated.

JAC.PhotoExcitation.computeCrossSection — Method

PhotoExcitation.computeCrossSection(line::PhotoExcitation.Line, stokes::ExpStokes) ... to compute the excitation cross section for the excitation of unpolarized atoms by plane-wave photons, whose polarization is described by the given (experimental) Stokes parameters. A cs::EmProperty is returned.

JAC.PhotoExcitation.computeLines — Method

PhotoExcitation.computeLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, grid::Radial.Grid, settings::PhotoExcitation.Settings; output=true) ... to compute the photo-excitation amplitudes and all properties as requested by the given settings. A list of lines::Array{PhotoExcitation.Lines} is returned.

JAC.PhotoExcitation.computeLinesCascade — Method

PhotoExcitation.computeLinesCascade(finalMultiplet::Multiplet, initialMultiplet::Multiplet, grid::Radial.Grid, settings::PhotoExcitation.Settings, initialLevelSelection::LevelSelection; output::Bool=true, printout::Bool=true) ... to compute the excitation (absorption) transition amplitudes and all properties as requested by the given settings. The computations and printout is adapted for larger cascade computations by including only lines with at least one channel and by sending all printout to a summary file only. A list of lines::Array{PhotoExcitation.Lines} is returned.

JAC.PhotoExcitation.computeStatisticalTensor — Method

PhotoExcitation.computeStatisticalTensor(k::Int64, q::Int64, line::PhotoExcitation.Line, stokes::ExpStokes) ... to compute the statistical tensor (component) rho{k,q} of the final level for the excitation of unpolarized atoms by plane-wave photons, whose polarization is described by the given (experimental) Stokes parameters. A rhokq::EmProperty is returned.

JAC.PhotoExcitation.determineChannels — Method

PhotoExcitation.determineChannels(finalLevel::Level, initialLevel::Level, settings::PhotoExcitation.Settings) ... to determine a list of PhotoExcitation.Channel for a transitions from the initial to final level and by taking into account the particular settings of for this computation; an Array{PhotoExcitation.Channel,1} is returned.

JAC.PhotoExcitation.determineLines — Method

PhotoExcitation.determineLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, settings::PhotoExcitation.Settings) ... to determine a list of photo-excitation Line's for transitions between the levels from the given initial- and final-state multiplets and by taking into account the particular selections and settings for this computation; an Array{PhotoExcitation.Line,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JAC.PhotoExcitation.displayCrossSections — Method

PhotoExcitation.displayCrossSections(stream::IO, lines::Array{PhotoExcitation.Line,1}, settings::PhotoExcitation.Settings) ... to list all results, energies, cross sections, etc. of the selected lines. A neat table is printed but nothing is returned otherwise.

JAC.PhotoExcitation.displayLineData — Method

PhotoExcitation.displayLineData(stream::IO, lines::Array{PhotoExcitation.Line,1}) ... to display the calculated data, ordered by the initial levels and the photon energies involved. Neat tables of all initial levels and photon energies as well as all associated total cross sections are printed but nothing is returned otherwise.

JAC.PhotoExcitation.displayLines — Method

PhotoExcitation.displayLines(stream::IO, lines::Array{PhotoExcitation.Line,1}) ... to display a list of lines and channels that have been selected due to the prior settings. A neat table of all selected transitions and energies is printed but nothing is returned otherwise.

JAC.PhotoExcitation.estimateCrossSection — Method

PhotoExcitation.estimateCrossSection(lines::Array{PhotoExcitation.Line,1}, omega::Float64, gamma::Float64, initialLevel) ... to estimate from lines the total PE cross section for any omega and for initial level. The procedure assumes a Gaussian line shape for each PhotoExcitation.Line with widths gamma and distributes the (total cross section according to the line shape within [Er - 10*gamma, Er + 10*gamma]. Other line shapes can be readily implemented; a cross section cs::EmProperty is returned.

JAC.PhotoIonization

JAC.PhotoIonization.amplitude — Method

PhotoIonization.amplitude(kind::String, channel::PhotoIonization.Channel, omega::Float64, continuumLevel::Level, initialLevel::Level, grid::Radial.Grid) ... to compute the kind = (photoionization) amplitude <(alphaf Jf, epsilon kappa) Jt || O^(photoionization) || alphai J_i> due to the electron-photon interaction for the given final and initial level, the partial wave of the outgoing electron as well as the given multipole and gauge. A value::ComplexF64 is returned.

JAC.PhotoIonization.angularFunctionK — Method

PhotoIonization.angularFunctionK(L1::Int64, L2::Int64, X::Int64, Ji::AngularJ64, Jf::AngularJ64, kappa1::Int64, J1::AngularJ64, kappa2::Int64, J2::AngularJ64) ... to compute angular function K(...) as defined for the non-E1 angle-differential cross sections by Nishita Hosea (2025). No tests are made that the triangular conditions of the quantum numbers are fulfilled. A wa::Float64 is returned.

JAC.PhotoIonization.angularFunctionW — Method

PhotoIonization.angularFunctionW(theta::Float64, L1::Int64, L2::Int64, X::Int64, lambda1::Int64, lambda2::Int64, kappa1::Int64, mu1::Rational{Int64}, kappa2::Int64, mu2::Rational{Int64}) ... to compute angular function W(theta; ...) as defined for the non-E1 angle-differential cross sections by Nishita Hosea (2025). No tests are made that the triangular conditions of the quantum numbers are fulfilled. A wa::Float64 is returned.

JAC.PhotoIonization.computeAmplitudesProperties — Method

PhotoIonization.computeAmplitudesProperties(line::PhotoIonization.Line, nm::Nuclear.Model, grid::Radial.Grid, nrContinuum::Int64, settings::PhotoIonization.Settings; printout::Bool=false) ... to compute all amplitudes and properties of the given line; a line::PhotoIonization.Line is returned for which the amplitudes and properties are now evaluated.

JAC.PhotoIonization.computeAmplitudesPropertiesPlasma — Method

PhotoIonization.computeAmplitudesPropertiesPlasma(line::PhotoIonization.Line, nm::Nuclear.Model, grid::Radial.Grid, settings::PhotoIonization.PlasmaSettings) ... to compute all amplitudes and properties of the given line but for the given plasma model; a line::PhotoIonization.Line is returned for which the amplitudes and properties are now evaluated.

JAC.PhotoIonization.computeAngularBeta — Method

PhotoIonization.computeAngularBeta(iLevel::Level, fLevel::Level, channels::Array{PhotoIonization.Channel,1}) ... to compute the beta anisotropy parameter for the photoionization transition i -> f with the given channels; here, the formula from Balashov (1994, Eq. 2.135) has been utilized. A beta::EmProperty parameter is returned. These (gauge-dependent) beta parameters are set to -9., if no amplitudes are calculated for the given gauge.

JAC.PhotoIonization.computeDisplayNonE1AngleDifferentialCS — Method

PhotoIonization.computeDisplayNonE1AngleDifferentialCS(stream::IO, lines::Array{PhotoIonization.Line,1}, settings::PhotoIonization.Settings) ... to compute & display the non-E1 angle-differential photoionization cross sections for all PhotoIonization.Line's and at all angles theta as defined in the settings. The general formula by Nishita Hosea (2025) is applied here. A neat table is printed for each line but nothing is returned otherwise.

JAC.PhotoIonization.computeLines — Method

PhotoIonization.computeLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::PhotoIonization.Settings; output::Bool=true) ... to compute the photoIonization transition amplitudes and all properties as requested by the given settings. A list of lines::Array{PhotoIonization.Lines} is returned.

JAC.PhotoIonization.computeLinesCascade — Method

PhotoIonization.computeLinesCascade(finalMultiplet::Multiplet, initialMultiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::PhotoIonization.Settings, initialLevelSelection::LevelSelection; output=true, printout::Bool=true) ... to compute the photoionization transition amplitudes and all properties as requested by the given settings. The computations and printout is adapted for large cascade computations by including only lines with at least one channel and by sending all printout to a summary file only. A list of lines::Array{PhotoIonization.Lines} is returned.

JAC.PhotoIonization.computeLinesPlasma — Method

PhotoIonization.computeLinesPlasma(finalMultiplet::Multiplet, initialMultiplet::Multiplet, nm::Nuclear.Model, grid::Radial.Grid, settings::PhotoIonization.PlasmaSettings; output::Bool=true) ... to compute the photoIonization transition amplitudes and all properties as requested by the given settings. A list of lines::Array{PhotoIonization.Lines} is returned.

JAC.PhotoIonization.computePartialCrossSectionUnpolarized — Method

PhotoIonization.computePartialCrossSectionUnpolarized(gauge::EmGauge, Mf::AngularM64, line::PhotoIonization.Line) ... to compute the partial photoionization cross section for initially unpolarized atoms by unpolarized plane-wave photons. A value::Float64 is returned.

JAC.PhotoIonization.computeStatisticalTensorUnpolarized — Method

PhotoIonization.computeStatisticalTensorUnpolarized(k::Int64, q::Int64, gauge::EmGauge, line::PhotoIonization.Line, settings::PhotoIonization.Settings) ... to compute the statistical tensor of the photoion in its final level after the photoionization of initially unpolarized atoms by plane-wave photons with given Stokes parameters (density matrix). A value::ComplexF64 is returned.

JAC.PhotoIonization.computeTimeDelays — Method

PhotoIonization.computeTimeDelays(channels::Array{PhotoIonization.Channel,1}, xchannels::Array{PhotoIonization.Channel,1}, deltaE::Float64, Jf::AngularJ64) ... to compute the – coherent and incoherent – time delay from the channels as calculated for two neighboured photon energies (deltaE = xE - E). Two tuple of two time delays (coherentDelay::EmProperty, incoherentDelay::EmProperty) is returned.

JAC.PhotoIonization.determineChannels — Method

PhotoIonization.determineChannels(finalLevel::Level, initialLevel::Level, settings::PhotoIonization.Settings) ... to determine a list of photoionization Channel for a transitions from the initial to final level and by taking into account the particular settings of for this computation; an Array{PhotoIonization.Channel,1} is returned.

JAC.PhotoIonization.determineLines — Method

PhotoIonization.determineLines(finalMultiplet::Multiplet, initialMultiplet::Multiplet, settings::PhotoIonization.Settings) ... to determine a list of PhotoIonization.Line's for transitions between levels from the initial- and final-state multiplets, and by taking into account the particular selections and settings for this computation; an Array{PhotoIonization.Line,1} is returned. Apart from the level specification, all physical properties are set to zero during the initialization process.

JAC.PhotoIonization.displayLineData — Method

PhotoIonization.displayLineData(stream::IO, lines::Array{PhotoIonization.Line,1}) ... to display the calculated data, ordered by the initial levels and the photon energies involved. Neat tables of all initial levels and photon energies as well as all associated cross sections are printed but nothing is returned otherwise.

JAC.PhotoIonization.displayLines — Method

PhotoIonization.displayLines(stream::IO, lines::Array{PhotoIonization.Line,1}) ... to display a list of lines and channels that have been selected due to the prior settings. A neat table of all selected transitions and energies is printed but nothing is returned otherwise.

JAC.PhotoIonization.displayPhases — Method

PhotoIonization.displayPhases(lines::Array{PhotoIonization.Line,1}) ... to display a list of lines, channels and phases of the continuum wave that have been selected due to the prior settings. A neat table of all selected transitions and energies is printed but nothing is returned otherwise.

JAC.PhotoIonization.displayResults — Method

PhotoIonization.displayResults(stream::IO, lines::Array{PhotoIonization.Line,1}, settings::PhotoIonization.Settings) ... to list all results, energies, cross sections, etc. of the selected lines. A neat table is printed but nothing is returned otherwise.

JAC.PhotoIonization.extractCrossSection — Method

PhotoIonization.extractCrossSection(lines::Array{PhotoIonization.Line,1}, omega::Float64, initialLevel) ... to extract from lines the total PI cross section that refer to the given omega and initial level; a cross section cs::EmProperty is returned.

JAC.PhotoIonization.extractCrossSection — Method

PhotoIonization.extractCrossSection(lines::Array{PhotoIonization.Line,1}, omega::Float64, shell::Shell, initialLevel) ... to extract from lines the total PI cross section that refer to the given omega and initial level and to to the ionization of an electron from shell; a cross section cs::EmProperty is returned.

JAC.PhotoIonization.extractLines — Method

PhotoIonization.extractLines(lines::Array{PhotoIonization.Line,1}, omega::Float64) ... to extract from lines all those that refer to the given omega; a reduced list rLines::Array{PhotoIonization.Line,1} is returned.

JAC.PhotoIonization.extractPhotonEnergies — Method

PhotoIonization.extractPhotonEnergies(lines::Array{PhotoIonization.Line,1}) ... to extract all photon energies for which photoionization data and cross sections are provided by lines; an list of energies::Array{Float64,1} is returned.

JAC.PhotoIonization.getLineKappas — Method

PhotoIonization.getLineKappas(line::PhotoIonization.Line) ... returns a list of kappa-values (partial waves) which contribute to the given line, to which one or several channels are assigned. An kappaList::Array{Int64,1} is returned.

JAC.PhotoIonization.interpolateCrossSection — Method

PhotoIonization.interpolateCrossSection(lines::Array{PhotoIonization.Line,1}, omega::Float64, initialLevel) ... to interpolate (or extrapolate) from lines the total PI cross section for any given omega and initial level. The procedure applies a linear interpolation/extrapolation by just using the cross sections from the two nearest (given) omega points; a cross section cs::EmProperty is returned.