Auto Ionization

JAC.AutoIonization.Channel — Type

struct Channel ... defines a type for a AutoIonization channel to help characterize a scattering (continuum) state of many electron-states with a single free electron.

+ kappa          ::Int64                ... partial-wave of the free electron
+ symmetry       ::LevelSymmetry        ... total angular momentum and parity of the scattering state
+ phase          ::Float64              ... phase of the partial wave
+ amplitude      ::Complex{Float64}     ... Auger amplitude associated with the given channel.

JAC.AutoIonization.Line — Type

struct Line ... defines a type for a AutoIonization line that may include the definition of sublines and their corresponding amplitudes.

+ initialLevel   ::Level           ... initial-(state) level
+ finalLevel     ::Level           ... final-(state) level
+ electronEnergy ::Float64         ... Energy of the (incoming free) electron.
+ totalRate      ::Float64         ... Total rate of this line.
+ angularAlpha   ::Float64         ... Angular alpha_2 coefficient.
+ channels       ::Array{AutoIonization.Channel,1}  ... List of AutoIonization channels of this line.

JAC.AutoIonization.Line — Method

AutoIonization.Line(initialLevel::Level, finalLevel::Level, totalRate::Float64) ... constructor for an AutoIonization line between a specified initial and final level.

JAC.AutoIonization.PlasmaSettings — Type

struct AutoIonization.PlasmaSettings <: Basics.AbstractLineShiftSettings ... defines a type for the details and parameters of computing Auger rates with plasma interactions.

+ printBefore         ::Bool             ... True, if all energies and lines are printed before their evaluation.
+ lineSelection       ::LineSelection    ... Specifies the selected levels, if any.

JAC.AutoIonization.PlasmaSettings — Method

AutoIonization.PlasmaSettings() ... constructor for a standard instance of AutoIonization.PlasmaSettings.

JAC.AutoIonization.Settings — Type

struct Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing Auger lines.

+ calcAnisotropy      ::Bool               ... True, if the intrinsic alpha_2,4 angular parameters are to be 
                                                calculated, and false otherwise.
+ calcTeAuger         ::Bool               
    ... True, if contributions of the two-electron Auger transitions are to be calculated, and false otherwise;
        this flag requires a proper (resonant) Green function that supports the TEA transitions.
+ printBefore         ::Bool               ... True, if all energies and lines are printed before their evaluation.
+ lineSelection       ::LineSelection      ... Specifies the selected levels, if any.
+ augerEnergyShift    ::Float64            ... An overall energy shift for all Auger (free-electron) energies.
+ minAugerEnergy      ::Float64            ... Minimum energy of free (Auger) electrons to be included.
+ maxAugerEnergy      ::Float64            ... Maximum energy of free (Auger) electrons to be included.
+ maxKappa            ::Int64              ... Maximum kappa value of partial waves to be included.
+ operator            ::AbstractEeInteraction   
    ... Auger operator that is to be used for evaluating the Auger amplitudes; allowed values are: 
        CoulombInteraction(), BreitInteraction(), ...
+ gMultiplet          ::Multiplet      
    ... Mean-field multiplet of intermediate levels in the computations, sometimes referred to as
        (resonant) Green function.

JAC.AutoIonization.Settings — Method

AutoIonization.Settings(set::AutoIonization.Settings;

    calcAnisotropy=..,      calcTeAuger..,              printBefore=..,         augerEnergyShift=.., 
    minAugerEnergy=..,      maxAugerEnergy=..,          maxKappa=..,            operator=..,
    gMultiplet=.. )
                
... constructor for modifying the given AutoIonization.Settings by 'overwriting' the previously selected parameters.

JAC.AutoIonization.Settings — Method

AutoIonization.Settings() ... constructor for the default values of AutoIonization line computations.

Dielectronic Recombination

JAC.DielectronicRecombination.AbstractCorrections — Type

abstract type DielectronicRecombination.AbstractCorrections ... defines an abstract type to distinguish different types of corrections to the decay rates and strength. These corrections are based on the classification of shell:

    n^(core)  <   n^(final)  <  n^(hydrogenic)  <  n^(lowest-captured)  <  n^(lower-empirical)    
              <=  n^(upper-empirical)              ... where
              
    n^(core)            ... refers to the (maximum) principal quantum number to which initial core electrons are excited;
    n^(final)           ... the maximum number for which shells are treated explicitly in the representation of the final levels f;
    n^(hydrogenic)      ... to the maximum n-shell, to which the radiative decay is modeled by scaled-hydrogenic rates, 
                        ... and which can be omitted also from the list. 
    n^(lowest-captured) ... is the lowest, high-n shell, into which the additional electron is captured and which must
                            (of course) occur explicitly in the basis of the intermediate and final levels. 
    [n^(lower-empirical)  <=  n^(upper-empirical)]  
                        ... designates additional (empirical) high-n shells for which the contributions to the DR resonances 
                            are still estimated empirically by using arguments from quantum-defect theory. All shells with 
                            n > n^(upper-empirical) are neglected completely for their contributions to the DR spectra; see also:

+ struct DielectronicRecombination.EmpiricalCorrections  
    ... to estimate empirically the contributions of additional resonances for the capture of an electron into
        shells with [n^(lower-empirical)  <=  n^(upper-empirical)]. A simple scaling of the rates, calculated initially
        for n^(lowest-captured), ... only, is utilized for estimating the associated strength for these additional
        resonances.
+ struct DielectronicRecombination.HydrogenicCorrections  
    ... to add for missing final decay levels to the (total) photon decay rates by scaling the corresponding rates
        of non-relativistic hydrogenic ions with a suitable effective charge (Zeff); these hydrogenic corrections improve
        goth, the total photon rate as well as the resonance strength.
+ struct DielectronicRecombination.MaximumlCorrection  
    ... to exclude all subshells with l > l_max in the hydrogenic corrections; this restriction does not apply to the 
        given resonance levels, which can be controlled (and are specified) by the list of intermediate configurations.

JAC.DielectronicRecombination.EmpiricalCorrections — Type

struct DielectronicRecombination.EmpiricalCorrections <: DielectronicRecombination.AbstractCorrections ... to include empirical corrections for the shells with [n^(lower-empirical) <= n^(upper-empirical)]. A rather rude model is used so far.

+ nUpperEmpirical ::Union{Int64,Missing}   
    ... The upper-empirical shell for which rate contributions are estimated; the lower-empirical shell = n^(captured-max + 1) 
        is derived from the given configuration lists. No corrections are made for nUpperEmpirical <= n^(captured-max + 1).
+ effectiveZ      ::Union{Float64,Missing}  ... effective charge Z_eff for the hydrogenic correction (inactive).
+ rateScaling     ::Union{Float64,Missing}  ... scaling factor to modify the estimated rates.

JAC.DielectronicRecombination.EmpiricalTreatment — Type

struct DielectronicRecombination.EmpiricalTreatment ... defines an (internal) type to communicate and distribute the physical (and technical) parameters that are utilized to make the requested empirical corrections or just nothing. This data type should not be applied by the user but is initialized by the given (set of) corrections. Otherwise, it is treated like any other type in JAC. All parameters are made physically "explicit", even if they were "missing" originally, and can be directly applied in the empirical treatment of the DR process. The following hierarchy of shells is used:

    n^(core)  <   n^(final)  <  n^(hydrogenic)  <  n^(lowest-captured)  <  n^(lower-empirical)    
              <=  n^(upper-empirical) 
    
+ doEmpiricalCorrections      ::Bool    ... True, if empirical corrections are needed, false o/w.
+ doHydrogenicCorrections     ::Bool    ... True, if hydrogenic corrections are needed, false o/w.
+ doMaximumlCorrection        ::Bool    ... True, if a maximum l values is used, false o/w.
+ doResonanceWindowCorrection ::Bool    ... True, if a window of resonances is specified, false o/w.
+ nCore                       ::Int64   
    ... (maximum) principal quantum number to which initial core electrons are excited;
+ nFinal                      ::Int64   
    ... the maximum number for which shells are treated explicitly in the representation of the final levels f;
+ nHydrogenic                 ::Int64   
    ... maximum n-shell, to which the radiative decay is modeled by scaled-hydrogenic rates. 
+ nLowestCaptured             ::Int64   
    ... lowest, high-n shell, into which the additional electron is captured and which must (of course) occur 
        explicitly in the basis of the intermediate and final levels.
+ nLowerEmpirical             ::Int64   
    ... maximum n-shell, to which the radiative decay is modeled by scaled-hydrogenic rates. 
+ nUpperEmpirical             ::Int64   
    ... additional (empirical) high-n shells for which the contributions to the DR resonances are still 
        estimated empirically by using arguments from quantum-defect theory.
+ maximum_l                   ::Int64    ... maximum l value; is set to a large value if not specified by the user.
+ hydrogenicEffectiveZ        ::Float64  ... effective charge Z_eff for the hydrogenic correction (inactive).
+ hydrogenicRateScaling       ::Float64  ... scaling factor to modify the estimated hydrogenic rates.
+ empiricalEffectiveZ         ::Float64  ... effective Z for empirical estimates
+ empiricalRateScaling        ::Float64  ... scaling factor to modify the empirical rates.
+ resonanceEnergyMin:         ::Float64  ... minimum energy [Hartree] of the resonances to be considered.
+ resonanceEnergyMax:         ::Float64  ... maximum energy [Hartree] of the resonances to be considered.

JAC.DielectronicRecombination.HydrogenicCorrections — Type

struct DielectronicRecombination.HydrogenicCorrections <: DielectronicRecombination.AbstractCorrections ... to add for missing final decay levels the photon decay rates for non-relativistic hydrogenic ions; this improves the total photon rate as well as the resonance strength. These corrections are taken into account for all shells with n^{final}+1 <= n <= nHydrogenic

+ nHydrogenic       ::Union{Int64,Missing}   
    ... upper principal quantum number nHydrogenic for which hydrogenic correctios to the radiative photon rates are 
        calculated explicitly; the photon rates are further scaled if some proper effectiveZ and/or rateScaling
        is provided.
+ effectiveZ      ::Union{Float64,Missing}   ... effective charge Z_eff for the hydrogenic correction.
+ rateScaling     ::Union{Float64,Missing}   ... scaling factor to scale the photon rates

JAC.DielectronicRecombination.MaximumlCorrection — Type

struct DielectronicRecombination.MaximumlCorrection <: DielectronicRecombination.AbstractCorrections ... to exclude all subshells with l > l_max, both in the treatment of the corrections shells.

+ maximum_l    ::Union{Int64,Missing}   
    ... maximum orbital angular momentum quantum number for which contributions to the DR strengths are 
        taken into account. This number applies for all subshells for which other corrections are 
        requested, whereas the "physical subshells" are defined by the configuration lists.

JAC.DielectronicRecombination.Passage — Type

struct DielectronicRecombination.Passage ... defines a type for a dielectronic recombination passage, i.e. a (reduced) pathways, that include the definition of channels and their corresponding amplitudes for the individual i –> m resonances, whereas the subsequent radiative stabilization is considered only later.

+ initialLevel      ::Level                   ... initial-(state) level
+ intermediateLevel ::Level                   ... intermediate-(state) level
+ electronEnergy    ::Float64                 ... energy of the (incoming, captured) electron
+ captureRate       ::Float64                 ... rate for the electron capture (Auger rate)
+ photonRate        ::EmProperty              ... rate for the photon emission
+ reducedStrength   ::EmProperty              
    ... reduced resonance strength Sum_f S(i -> d -> f) * Gamma_d of this passage; this reduced strength does 
        not require the knowledge of Gamma_d for the individual passage.
+ captureChannels   ::Array{AutoIonization.Channel,1}   ... List of |i> -->  |n>   dielectronic (Auger) capture channels.

JAC.DielectronicRecombination.Passage — Method

DielectronicRecombination.Passage() ... constructor for an 'empty' instance of a dielectronic recombination passage between a specified initial and intermediate level.

JAC.DielectronicRecombination.Pathway — Type

struct DielectronicRecombination.Pathway ... defines a type for a dielectronic recombination pathways that may include the definition of channels and their corresponding amplitudes.

+ initialLevel      ::Level                   ... initial-(state) level
+ intermediateLevel ::Level                   ... intermediate-(state) level
+ finalLevel        ::Level                   ... final-(state) level
+ electronEnergy    ::Float64                 ... energy of the (incoming, captured) electron
+ photonEnergy      ::Float64                 ... energy of the (emitted) photon
+ captureRate       ::Float64                 ... rate for the electron capture (Auger rate)
+ photonRate        ::EmProperty              ... rate for the photon emission
+ angularBeta       ::EmProperty              ... beta parameter of the photon emission
+ reducedStrength   ::EmProperty              ... reduced resonance strength S(i -> d -> f) * Gamma_d of this pathway;
                                                    this reduced strength does not require the knowledge of Gamma_d for each pathway.
+ captureChannels   ::Array{AutoIonization.Channel,1}   ... List of |i> -->  |n>   dielectronic (Auger) capture channels.
+ photonChannels    ::Array{PhotoEmission.Channel,1}    ... List of |n> -->  |f>   radiative stabilization channels.

JAC.DielectronicRecombination.Pathway — Method

DielectronicRecombination.Pathway() ... constructor for an 'empty' instance of a dielectronic recombination pathway between a specified initial, intermediate and final level.

JAC.DielectronicRecombination.Resonance — Type

struct DielectronicRecombination.Resonance ... defines a type for a dielectronic resonance as defined by a given initial and resonance level but by summing over all final levels

+ initialLevel      ::Level             ... initial-(state) level
+ intermediateLevel ::Level             ... intermediate-(state) level
+ resonanceEnergy   ::Float64           ... energy of the resonance w.r.t. the inital-state
+ resonanceStrength ::EmProperty        ... strength of this resonance due to the stabilization into any of the allowed final levels.
+ captureRate       ::Float64           ... capture (Auger) rate to form the intermediate resonance, starting from the initial level.
+ augerRate         ::Float64           ... total (Auger) rate for an electron emission of the intermediate resonance
+ photonRate        ::EmProperty        ... total photon rate for a photon emission, i.e. for stabilization.

JAC.DielectronicRecombination.Resonance — Method

DielectronicRecombination.Resonance() ... constructor for an 'empty' instance of a dielectronic resonance as defined by a given initial and resonance level but by summing over all final levels.

JAC.DielectronicRecombination.ResonanceSelection — Type

struct DielectronicRecombination.ResonanceSelection ... defines a type for selecting classes of resonances in terms of leading configurations.

+ active          ::Bool              ... initial-(state) level
+ fromShells      ::Array{Shell,1}    ... List of shells from which excitations are to be considered.
+ toShells        ::Array{Shell,1}    ... List of shells to which (core-shell) excitations are to be considered.
+ intoShells      ::Array{Shell,1}    ... List of shells into which electrons are initially placed (captured).

JAC.DielectronicRecombination.ResonanceSelection — Method

DielectronicRecombination.ResonanceSelection() ... constructor for an 'empty' instance of a ResonanceSelection()

JAC.DielectronicRecombination.ResonanceWindowCorrection — Type

struct DielectronicRecombination.ResonanceWindowCorrection <: DielectronicRecombination.AbstractCorrections ... to exclude all DR resonances outside of a given "window [Emin, Emax]" of resonance energies with regard to the initial level.

+ energyMin  ::Float64   ... minimum energy [Hartree] of the resonances to be considered.  
+ energyMax  ::Float64   ... maximum energy [Hartree] of the resonances to be considered.

JAC.DielectronicRecombination.Settings — Type

struct DielectronicRecombination.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing dielectronic recombination pathways.

+ multipoles            ::Array{EmMultipoles}  ... Multipoles of the radiation field that are to be included.
+ gauges                ::Array{UseGauge}      ... Specifies the gauges to be included into the computations.
+ calcOnlyPassages      ::Bool                 
    ... Only compute resonance strength but without making all the pathways explicit. This option is useful
        for the capture into high-n shells or if the photons are not considered explicit. It also treats the 
        shells differently due to the given core shells < final-state shells < hydrogenically-scaled shells <
        capture-shells < asymptotic-shells. Various correction and multi-threading techiques can be applied
        to deal with or omit different classes of these shells.
+ calcRateAlpha         ::Bool                 
    ... True, if the DR rate coefficients are to be calculated, and false o/w.
+ printBefore           ::Bool                 
    ... True, if all energies and pathways are printed before their evaluation.
+ pathwaySelection      ::PathwaySelection     ... Specifies the selected levels/pathways, if any.
+ electronEnergyShift   ::Float64              
    ... An overall energy shift for all electron energies (i.e. from the initial to the resonance levels [Hartree].
+ photonEnergyShift     ::Float64              
    ... An overall energy shift for all photon energies (i.e. from the resonance to the final levels.
+ mimimumPhotonEnergy   ::Float64              
    ... minimum transition energy for which photon transitions are  included into the evaluation.
+ temperatures          ::Array{Float64,1}     
    ... list of temperatures for which plasma rate coefficients are displayed; however, these rate coefficients
        only include the contributions from those pathsways that are calculated here explicitly.
+ corrections           ::Array{DielectronicRecombination.AbstractCorrections,1}
    ... Specify, if appropriate, the inclusion of additional corrections to the rates and DR strengths.
+ augerOperator         ::AbstractEeInteraction 
    ... Auger operator that is to be used for evaluating the Auger amplitude's; the allowed values are: 
        CoulombInteraction(), BreitInteration(), CoulombBreit(), CoulombGaunt().

JAC.DielectronicRecombination.Settings — Method

(set::DielectronicRecombination.Settings;

    multipoles=..,             gauges=..,                  
    calcOnlyPassages=..,       calcRateAlpha=..,         printBefore=..,           pathwaySelection=..,     
    electronEnergyShift=..,    photonEnergyShift=..,       
    mimimumPhotonEnergy=..,    temperatures=..,          corrections=..,           augerOperator=..)
                
... constructor for modifying the given DielectronicRecombination.Settings by 'overwriting' the previously selected parameters.

JAC.DielectronicRecombination.Settings — Method

DielectronicRecombination.Settings() ... constructor for the default values of dielectronic recombination pathway computations.

Electron Impact Excitation

JAC.ImpactExcitation.Channel — Type

struct ImpactExcitation.Channel ... defines a type for a electron-impact excitaiton channel to help characterize the incoming and outgoing (continuum) states of many electron-states with a single free electron

+ initialKappa     ::Int64              ... partial-wave of the incoming free electron
+ finalKappa       ::Int64              ... partial-wave of the outgoing free electron
+ symmetry         ::LevelSymmetry      ... total angular momentum and parity of the scattering state
+ initialPhase     ::Float64            ... phase of the incoming partial wave
+ finalPhase       ::Float64            ... phase of the outgoing partial wave
+ amplitude        ::Complex{Float64}   ... Collision amplitude associated with the given channel.

JAC.ImpactExcitation.Line — Type

struct ImpactExcitation.Line ... defines a type for a electron-impact excitation line that may include the definition of channels and their corresponding amplitudes.

+ initialLevel           ::Level         ... initial- (bound-state) level
+ finalLevel             ::Level         ... final- (bound-state) level
+ initialElectronEnergy  ::Float64       ... energy of the incoming (initial-state) free-electron
+ finalElectronEnergy    ::Float64       ... energy of the outgoing (final-state) free-electron
+ crossSection           ::Float64       ... total cross section of this line
+ collisionStrength      ::Float64       ... total collision strength of this line
+ channels               ::Array{ImpactExcitation.Channel,1}  ... List of ImpactExcitation channels of this line.
+ convergence            ::Float64       ... convergence of calculation

JAC.ImpactExcitation.Line — Method

ImpactExcitation.Line(initialLevel::Level, finalLevel::Level, crossSection::Float64) ... constructor for an electron-impact excitation line between a specified initial and final level.

JAC.ImpactExcitation.Line — Method

ImpactExcitation.Line() ... 'empty' constructor for an electron-impact excitation line between a specified initial and final level.

JAC.ImpactExcitation.RateCoefficients — Type

struct ImpactExcitation.RateCoefficients ... Defines a type for the output results from excitation rate or effective collision strengths calculations + initialLevel ::Level ... initial- (bound-state) level + finalLevel ::Level ... final- (bound-state) level + temperatures ::Array{Float64,1} ... Temperatures in [K] to calculate excitation rates and effective collision strengths + alphas ::Array{Float64,1} ... Excitation rate coefficients in [cm^3/s] for the input temperatures + effOmegas ::Array{Float64,1} ... Effective collision strengths for the input temperatures

JAC.ImpactExcitation.RateCoefficients — Method

ImpactExcitation.RateCoefficients() ... constructor for the default values of electron-impact excitation rate coefficient or effective collision strength computations.

JAC.ImpactExcitation.Settings — Type

struct ImpactExcitation.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing electron-impact excitation lines. + lineSelection ::LineSelection ... Specifies the selected levels, if any. + electronEnergies ::Array{Float64,1} ... List of impact-energies of the incoming elecgtrons (in user-defined units). + energyShift ::Float64 ... An overall energy shift for all transitions |i> –> |f>. + maxKappa ::Int64 ... Maximum kappa value of partial waves to be included. + calcRateCoefficient ::Bool ... True, if the plasma rate coefficients to be calculated, false otherwise. + maxEnergyMultiplier ::Float64 ... Maximum initial electron energy for eff. collision strength integration. (maxEnergyMultiplier * Excitation threshold energy), after this assymptotic limit is applied. + numElectronEnergies ::Int64 ... No. of different electron energy points at which collision strengths to compute in electron energy range [0, (maxEnergyMultiplier * Excitation threshold energy)] + temperatures ::Array{Float64, 1} ... Electron temperatures [K] for eff. collision strengths and rate coefficients. + printBefore ::Bool ... True, if all energies and lines are printed before their evaluation. + operator ::AbstractEeInteraction ... Interaction operator that is to be used for evaluating the e-e interaction amplitudes; allowed values are: CoulombInteraction(), BreitInteraction(), ...

JAC.ImpactExcitation.Settings — Method

ImpactExcitation.Settings(set::ImpactExcitation.Settings;

lineSelection...,  electronEnergies..., energyShift..., maxKappa...,
calcRateCoefficient..., maxEnergyMultiplier..., numElectronEnergies..., temperatures...,
printBefore..., operator...)

... constructor for modifying the given ImpactExcitation.Settings by 'overwriting' the previously selected parameters.

JAC.ImpactExcitation.Settings — Method

ImpactExcitation.Settings() ... constructor for the default values of electron-impact excitation line computations.

Photo Emission

JAC.PhotoEmission.Channel — Type

struct PhotoEmission.Channel ... defines a type for a single radiative emission/absorption channel that specifies the multipole, gauge and amplitude.

+ multipole         ::EmMultipole        ... Multipole of the photon emission/absorption.
+ gauge             ::EmGauge            ... Gauge for dealing with the (coupled) radiation field.
+ amplitude         ::Complex{Float64}   ... Amplitude of this multiple channel.

JAC.PhotoEmission.Line — Type

struct PhotoEmission.Line ... defines a type for a radiative line that may include the definition of sublines and their corresponding amplitudes.

+ initialLevel   ::Level               ... initial-(state) level
+ finalLevel     ::Level               ... final-(state) level
+ omega          ::Float64             ... Transition frequency of this line; can be shifted w.r.t. the level energies.
+ photonRate     ::EmProperty          ... Total rate of this line.
+ angularBeta    ::EmProperty          ... Angular beta_2 coefficient.
+ hasSublines    ::Bool                ... Determines whether the sublines are defined in terms of their multipolarity, amplitude, or not.
+ channels       ::Array{PhotoEmission.Channel,1}  ... List of radiative (photon) channels

JAC.PhotoEmission.Line — Method

PhotoEmission.Line(initialLevel::Level, finalLevel::Level, photonRate::Float64) ... constructor an radiative line between a specified initial and final level.

JAC.PhotoEmission.Settings — Type

struct PhotoEmission.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing radiative lines.

+ multipoles              ::Array{EmMultipoles}     ... Specifies the (radiat. field) multipoles to be included.
+ gauges                  ::Array{UseGauge}         ... Gauges to be included into the computations.
+ calcAnisotropy          ::Bool                    ... True, if the anisotropy (structure) functions are to be 
                                                        calculated and false otherwise 
+ printBefore             ::Bool                    ... True, if all energies and lines are printed before comput.
+ corePolarization        ::CorePolarization        ... Parametrization of the core-polarization potential/contribution.
+ lineSelection           ::LineSelection           ... Specifies the selected levels, if any.
+ photonEnergyShift       ::Float64                 ... An overall energy shift for all photon energies.
+ mimimumPhotonEnergy     ::Float64                 ... minimum transition energy for which (photon) transitions 
                                                        are included into the computation.
+ maximumPhotonEnergy     ::Float64                 ... maximum transition energy for which (photon) transitions 
                                                        are included.

JAC.PhotoEmission.Settings — Method

PhotoEmission.Settings(set::PhotoEmission.Settings;

    multipoles::=..,        gauges=..,                calcAnisotropy=..,          printBefore=..,
    corePolarization=..,    lineSelection=..,         photonEnergyShift=..,       
    mimimumPhotonEnergy=.., maximumPhotonEnergy=..) 
                
... constructor for modifying the given PhotoEmission.Settings by 'overwriting' the previously selected parameters.

JAC.PhotoEmission.Settings — Method

PhotoEmission.Settings() ... constructor for the default values of radiative line computations

Photo Excitation

JAC.PhotoExcitation.Line — Type

struct PhotoExcitation.Line ... defines a type for a photo-excitation line that may include the definition of sublines and their corresponding amplitudes.

+ initialLevel   ::Level                       ... initial-(state) level
+ finalLevel     ::Level                       ... final-(state) level
+ omega          ::Float64                     ... Transition frequency of this line; can be shifted w.r.t. the level energies.
+ oscStrength    ::EmProperty                  ... Absorption oscillator strength
+ crossSection   ::EmProperty                  ... Total cross section of this line.
+ staTensor      ::Array{TensorComp,1}         ... Array of statistical tensor components rho_kq
+ hasSublines    ::Bool                        ... Determines whether the individual sublines are defined in terms of their 
                                                    multipolarity, amplitude, or not; cf. PhotoEmission.Channel
+ channels       ::Array{PhotoEmission.Channel,1}  ... List of radiative (photon) channels

JAC.PhotoExcitation.Line — Method

PhotoExcitation.Line(initialLevel::Level, finalLevel::Level, omega::Float64, crossSection::EmProperty) ... constructor an photo-excitation line between a specified initial and final level.

JAC.PhotoExcitation.Settings — Type

struct PhotoExcitation.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing photo-excitation lines.

+ multipoles              ::Array{EmMultipole,1}    ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                  ::Array{UseGauge,1}       ... Specifies the gauges to be included into the computations.
+ calcForStokes           ::Bool                    ... True, if the excitation cross sections are to be calculated (and false otherwise)
                                                        for given Stokes parameter of the incident plane-wave photons.
+ calcPhotonDm            ::Bool                    ... True, if the photon density matrix of a subsequently emitted fluorescence photon 
                                                        is to be calculated and false otherwise. 
+ calcTensors             ::Bool                    ... True, if statistical tensors of the excited atom are to be calculated, false otherwise. 
+ printBefore             ::Bool                    ... True, if all energies and lines are printed before their evaluation.
+ lineSelection           ::LineSelection           ... Specifies the selected levels, if any.
+ photonEnergyShift       ::Float64                 ... An overall energy shift for all photon energies.
+ mimimumPhotonEnergy     ::Float64                 ... minimum transition energy for which (photon) transitions are included into the
                                                        computation.
+ maximumPhotonEnergy     ::Float64                 ... maximum transition energy for which (photon) transitions are included.
+ stokes                  ::ExpStokes               ... Stokes parameters of the incident radiation.

JAC.PhotoExcitation.Settings — Method

PhotoExcitation.Settings(set::PhotoExcitation.Settings;

    multipoles=..,          gauges=..,                  calcForStokes=..,           calcPhotonDm=..,    
    calcTensors=..,         printBefore=..,             lineSelection=..,    
    photonEnergyShift=..,   mimimumPhotonEnergy=..,     maximumPhotonEnergy=..,     stokes=..)
                
... constructor for modifying the given PhotoExcitation.Settings by 'overwriting' the previously selected parameters.

JAC.PhotoExcitation.Settings — Method

PhotoExcitation.Settings() ... 'empty' constructor for the default values of photo-excitation line computations

Photo Ionization

JAC.PhotoIonization.Channel — Type

struct PhotoIonization.Channel ... defines a type for a photoionization channel to help characterize a single multipole and scattering (continuum) state of many electron-states with a single free electron.

+ multipole      ::EmMultipole          ... Multipole of the photon absorption.
+ gauge          ::EmGauge              ... Gauge for dealing with the (coupled) radiation field.
+ kappa          ::Int64                ... partial-wave of the free electron
+ symmetry       ::LevelSymmetry        ... total angular momentum and parity of the scattering state
+ phase          ::Float64              ... phase of the partial wave
+ amplitude      ::Complex{Float64}     ... Photoionization amplitude associated with the given channel.

JAC.PhotoIonization.Line — Type

struct Line ... defines a type for a photoionization line that may include the definition of channels.

+ initialLevel   ::Level                  ... initial-(state) level
+ finalLevel     ::Level                  ... final-(state) level
+ electronEnergy ::Float64                ... Energy of the (outgoing free) electron.
+ photonEnergy   ::Float64                ... Energy of the absorbed photon.
+ crossSection   ::EmProperty             ... Cross section for this photoionization.
+ angularBeta    ::EmProperty             ... beta -parameter for unpolarized targets with J=0, 1/2, 1
+ coherentDelay  ::EmProperty             ... coherent time-delay due to the selected averaging of phases.
+ incoherentDelay::EmProperty             ... incoherent time-delay due to the selected averaging of phases.
+ channels       ::Array{PhotoIonization.Channel,1}  ... List of PhotoIonization.Channels of this line.

JAC.PhotoIonization.Line — Method

PhotoIonization.Line(initialLevel::Level, finalLevel::Level, crossSection::Float64) ... constructor for an photoionization line between a specified initial and final level.

JAC.PhotoIonization.PlasmaSettings — Type

struct PhotoIonization.PlasmaSettings <: Basics.AbstractLineShiftSettings ... defines a type for the details and parameters of computing photoionization rates with plasma interactions.

+ multipoles             ::Array{Basics.EmMultipole}     ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                 ::Array{Basics.UseGauge}        ... Specifies the gauges to be included into the computations.
+ photonEnergies         ::Array{Float64,1}              ... List of photon energies.  
+ printBefore            ::Bool                          ... True, if all energies and lines are printed before their evaluation.
+ lineSelection          ::LineSelection                 ... Specifies the selected levels, if any.

JAC.PhotoIonization.PlasmaSettings — Method

PhotoIonization.PlasmaSettings() ... constructor for a standard instance of PhotoIonization.PlasmaSettings.

JAC.PhotoIonization.Settings — Type

struct PhotoIonization.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing photoionization lines.

+ multipoles                    ::Array{EmMultipole}  ... Specifies the multipoles of the radiation field that are to be included.
+ gauges                        ::Array{UseGauge}     ... Specifies the gauges to be included into the computations.
+ photonEnergies                ::Array{Float64,1}    ... List of photon energies [in user-selected units].  
+ electronEnergies              ::Array{Float64,1}    ... List of electron energies; usually only one of these lists are utilized. 
+ thetas                        ::Array{Float64,1}    ... List of theta-values if angle-differential CS are calculated explicitly. 
+ calcAnisotropy                ::Bool                ... True, if the beta anisotropy parameters are to be calculated and false otherwise (o/w). 
+ calcPartialCs                 ::Bool                ... True, if partial cross sections are to be calculated and false otherwise.  
+ calcTimeDelay                 ::Bool                ... True, if time-delays are to be calculated and false otherwise.  
+ calcNonE1AngleDifferentialCS  ::Bool                ... True, if non-E1 angle-differential CS are be calculated and false otherwise.  
+ calcTensors                   ::Bool                ... True, if statistical tensors of the excited atom are to be calculated and false o/w. 
+ printBefore                   ::Bool                ... True, if all energies and lines are printed before their evaluation.
+ lineSelection                 ::LineSelection       ... Specifies the selected levels, if any.
+ stokes                        ::ExpStokes           ... Stokes parameters of the incident radiation.
+ freeElectronShift             ::Float64             ... An overall energy shift of all free-electron energies [user-specified units].
+ lValues                       ::Array{Int64,1}      ... Orbital angular momentum of free-electrons, for which partial waves are considered.

JAC.PhotoIonization.Settings — Method

PhotoIonization.Settings(set::PhotoIonization.Settings;

    multipoles=..,                      gauges=..,                  photonEnergies=..,          electronEnergies=..,     
    thetas=..,                          calcAnisotropy=..,          calcPartialCs..,            calcTimeDelay=..,           
    calcNonE1AngleDifferentialCS=..,    calcTensors=..,             printBefore=..,             lineSelection=..,           
    stokes=..,                          freeElectronShift=..,       lValues=.. )
                
... constructor for modifying the given PhotoIonization.Settings by 'overwriting' the previously selected parameters.

JAC.PhotoIonization.Settings — Method

PhotoIonization.Settings() ... constructor for the default values of photoionization line computations

Photo Recombination

JAC.PhotoRecombination.Channel — Type

struct PhotoRecombination.Channel ... defines a type for a photorecombination channel to help characterize a single multipole and scattering (continuum) state of many electron-states with a single free electron.

+ multipole      ::EmMultipole          ... Multipole of the photon emission/absorption.
+ gauge          ::EmGauge              ... Gauge for dealing with the (coupled) radiation field.
+ kappa          ::Int64                ... partial-wave of the free electron
+ symmetry       ::LevelSymmetry        ... total angular momentum and parity of the scattering state
+ phase          ::Float64              ... phase of the partial wave
+ amplitude      ::Complex{Float64}     ... Rec amplitude associated with the given channel.

JAC.PhotoRecombination.Line — Type

struct PhotoRecombination.Line ... defines a type for a Photorecombination line that may include the definition of channels.

+ initialLevel   ::Level                  ... initial-(state) level
+ finalLevel     ::Level                  ... final-(state) level
+ electronEnergy ::Float64                ... Energy of the (incoming free) electron.
+ photonEnergy   ::Float64                ... Energy of the emitted photon.
+ betaGamma2     ::Float64                ... beta^2 * gamma^2.
+ weight         ::Float64                ... weight of line in the integration over electron energies.
+ crossSection   ::EmProperty             ... Cross section for this electron capture.
+ channels       ::Array{PhotoRecombination.Channel,1}    ... List of photorecombination channels of this line.

JAC.PhotoRecombination.Line — Method

PhotoRecombination.Line() ... constructor for an empty instance of a photorecombination line between a specified initial and final level.

JAC.PhotoRecombination.Settings — Type

struct PhotoRecombination.Settings <: AbstractProcessSettings ... defines a type for the details and parameters of computing photo recombination lines.

+ multipoles          ::Array{EmMultipole}  ... Multipoles of the radiation field that are to be included.
+ gauges              ::Array{UseGauge}     ... Gauges to be included into the computations.
+ electronEnergies    ::Array{Float64,1}    ... List of electron energies [in default units].
+ ionEnergies         ::Array{Float64,1}    ... List of ion energies [in MeV/u].
+ useIonEnergies      ::Bool                ... Make use of ion energies in [MeV/u] to obtain the electron energies.
+ calcTotalCs         ::Bool                ... True, if the total cross sections is to be calculated/displayed for all initial levels.
+ calcAnisotropy      ::Bool                ... True, if the overall anisotropy is to be calculated.
+ calcTensors         ::Bool                ... True, if the statistical tensors are to be calculated and 
                                                false otherwise.
+ printBefore         ::Bool                ... True, if all energies and lines are printed before their evaluation.
+ maxKappa            ::Int64               ... Maximum kappa value of partial waves to be included.
+ lineSelection       ::LineSelection       ... Specifies the selected levels, if any.

JAC.PhotoRecombination.Settings — Method

PhotoRecombination.Settings(set::PhotoRecombination..Settings;

    multipoles=..,          gauges=..,              electronEnergies=..,          ionEnergies=..,     
    useIonEnergies=..,      calcTotalCs..,          calcAnisotropy=..,            calcTensors=..,             
    printBefore=..,         maxKappa=..,            lineSelection=..)
                
... constructor for modifying the given PhotoRecombination..Settings by 'overwriting' the previously selected parameters.

JAC.PhotoRecombination.Settings — Method

PhotoRecombination.Settings() ... constructor for the default values of photo recombination line computations