Empirical computations

JAC.Empirical.Computation — Type

struct Empirical.Computation ... defines a type for an empirical computation of various ionization and charge exchange processes.

+ name                   ::String                           ... A name associated to the computation.
+ nuclearModel           ::Nuclear.Model                    ... Model, charge and parameters of the nucleus.
+ grid                   ::Radial.Grid                      ... The radial grid to be used for the computation.
+ configs                ::Array{Configuration,1}           ... A list of non-relativistic configurations.
+ settings               ::Basics.AbstractEmpiricalSettings ... Provides the settings for the selected computations.

JAC.Empirical.Computation — Method

Empirical.Computation(comp::Empirical.Computation;

name=..,                nuclearModel=..,            grid=..,                    configs=..,                   settings=..,  
printout::Bool=false)
                
... constructor for modifying the given Empirical.Computation by 'overwriting' the previously selected parameters.

JAC.Empirical.Computation — Method

Empirical.Computation() ... constructor for an 'empty' instance::Empirical.Computation.

JAC.Basics.perform — Method

Basics.perform(computation::Empirical.Computation) ... to set-up and perform an empirical computation that starts from a given nuclear model and set of configurations, and which is mainly controlled by its settings. The results are printed to screen but nothing is returned otherwise.

Basics.perform(computation::Empirical.Computation; output=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the kind and specifications of the empirical computation but can easily accessed by the keys of this dictionary.

Semi-empirical estimates

JAC.Semiempirical.AtomicCrossSection — Type

@enum AtomicCrossSection ... defines a enumeration for (supported) cross section estimation.

JAC.Semiempirical.AtomicCrossSection — Method

Semiempirical.AtomicCrossSection(sa::String) ... constructor for a given String.

JAC.Semiempirical.estimate — Method

Semiempirical.estimate("binding energy: ...", Z::Float64, conf::Configuration) ... to provide an approximate binding energy of a given electron configuration. This estimate adds the binding energies of all subshell, taken frogm a semi-empirical tabulations by Williams et al., https://userweb.jlab.org/~gwyn/ebindene.html. If no binding energy is provided by this table, it simply scales the binding energies (e/ 12.) from the next lower shell of the same symmetry No relaxation effects are included if several hole states occur with regard to the neutral atom. An energy::Float64 in Hartree is returned.

JAC.Semiempirical.estimate — Method

Semiempirical.estimate("binding energy", Z::Int64, sh::Subshell; useLarkins::Bool=false) ... to provide the binding energy of a subshell electron, taken from a semi-empirical tabulations by Williams et al. (2000), https://userweb.jlab.org/~gwyn/ebindene.html or Larkins (1977). A energy::Float64 in Hartree is returned.

JAC.Semiempirical.estimate — Method

Semiempirical.estimate("ionization potential: inner-shell", shell::Shell, Z::Int64) ... to estimate the ionization potential = mean binding energy of an electron in the given shell; an energy value::Float64 is returned. These ionization potentials are taken from the tabulation ....

JAC.Semiempirical.estimateBindingEnergies — Method

Semiempirical.estimateBindingEnergies(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}, l::Int64) ... to estimate the binding energies of the high-n shells with n = nRange and orbital angular momentum l for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration. A list of energies::Array{Float64} [in Hartree] is returned.

JAC.Semiempirical.estimateBindingEnergies — Method

Semiempirical.estimateBindingEnergies(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}) ... to estimate the binding energies of the high-n shells with n = nRange for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration. A list of energies::Array{Float64} [in Hartree] is returned.

JAC.Semiempirical.estimateRadialExpectation — Method

Semiempirical.estimateRadialExpectation(Z::Float64, coreConf::Configuration, nRange::UnitRange{Int64}, l::Int64) ... to estimate the radial extent (r-expectation value) of the high-n shells with n = nRange and orbital angular momentum l for a (multiply-charged) ion with nuclear charge Z and core configuration coreConfiguration. A list of radii::Array{Float64} [in a_o] is returned.