Plasma computations

JAC.Plasma.AbstractPlasmaScheme — Type

abstract type Plasma.AbstractPlasmaScheme ... defines an abstract type to distinguish different kinds of plasma computations; see also:

+ struct AverageAtomScheme    
    ... to perform an average-atom computation.
+ struct LineShiftScheme    
    ... to compute the energy shifts and properties of atomic/ionic lines in some selected plasma model.
+ struct RateSchemePhotoionization    
    ... to compute (empirical) photoionization plasma rates and rate coefficients.
+ struct RateSchemePhotorecombination    
    ... to compute (empirical) photorecombination plasma rates and rate coefficients.
+ struct SahaBoltzmannScheme    
    ... to compute thermodynamic properties of a Saha-Boltzmann LTE mixture.

JAC.Plasma.AverageAtomScheme — Type

struct Plasma.AverageAtomScheme <: Plasma.AbstractPlasmaScheme ... a struct to perform an average-atom computation.

+ nMax                  ::Int64                ... maximum principal quantum number for the subshells of the AA model.
+ lMax                  ::Int64                ... maximum orbital quantum number for the subshells of the AA model.
+ scField               ::AbstractScField      ... maximum orbital quantum number for the subshells of the AA model.
+ calcPhotoionizationCs ::Bool                 ... True, if photoionization cross sections are to be calculated.
+ calcFormFactor        ::Bool                 ... True, if the form factor need to be calculated.
+ calcScatteringFactor  ::Bool                 ... True, if scattering factor need to be calculated.
+ piSubshells           ::Array{Subshell,1}    ... Bound subshells to be included into the photoionization cross sections.
+ omegas                ::Array{Float64,1}     ... energies for the photoionization & scattering factors [in a.u.].
+ qValues               ::Array{Float64,1}     ... q-values for calculating the form factor [in a.u.].

JAC.Plasma.AverageAtomScheme — Method

Plasma.AverageAtomScheme() ... constructor for an 'default' instance of a Plasma.AverageAtomScheme.

JAC.Plasma.Computation — Type

struct Computation ... defines a type for defining (simple) plasma computation for atoms and ions in a given set of reference configurations. It also support different atomic processes under plasma conditions. The plasma environment is typical described in terms of its temperature, density, etc.

+ scheme                         ::AbstractPlasmaScheme            ... Scheme (kind) of plasma computation.
+ nuclearModel                   ::Nuclear.Model                   ... Model, charge and parameters of the nucleus.
+ grid                           ::Radial.Grid                     ... The radial grid to be used for the computation.
+ refConfigs                     ::Array{Configuration,1}          ... A list of non-relativistic configurations.
+ asfSettings                    ::AsfSettings                     
    ... Provides the settings for the SCF process (under plasma conditions) and the associated CI calculations.
+ settings                       ::Plasma.Settings                 ... communicates the properties of the plasma

JAC.Plasma.Computation — Method

Plasma.Computation( ... example for plasma SCF computations)

    grid     = Radial.Grid(true)
    nuclearM = Nuclear.Model(18., "Fermi")
    ...
    refConfigs  = [Configuration("[Ne] 3s^2 3p^5")]
    Plasma.Computation(Plasma.Computation(), grid=grid, nuclearModel=nuclearM, refConfigs=refConfigs, asfSettings=... )

... These simple examples can be further improved by overwriting the corresponding parameters.

JAC.Plasma.Computation — Method

Plasma.Computation(comp::Plasma.Computation;

scheme=..,                  nuclearModel=..,            grid=..,                refConfigs=..,              asfSettings=..,     
settings=..,
printout::Bool=false)
                
... constructor for modifying the given Plasma.Computation by 'overwriting' the previously selected parameters.

JAC.Plasma.Computation — Method

Plasma.Computation() ... constructor for an 'empty' instance::Plasma.Computation.

JAC.Plasma.LineShiftScheme — Type

struct Plasma.LineShiftScheme <: Plasma.AbstractPlasmaScheme ... defines a type for the details and parameters of computing level energies with plasma interactions.

+ plasmaModel      ::AbstractPlasmaModel          ... Specify a particular plasma model, e.g. ion-sphere, Debye.
+ initialConfigs   ::Array{Configuration,1}       ... List of one or several configurations that define the initial-state multiplet.
+ finalConfigs     ::Array{Configuration,1}       ... List of one or several configurations that define the final-state multiplet.
+ settings         ::AbstractLineShiftSettings    ... Specify the process and settings for which line-shifts need to be computed.
## + NoBoundElectrons ::Int64                      ... Effective number of bound electrons.

JAC.Plasma.LineShiftScheme — Method

Plasma.LineShiftScheme() ... constructor for a standard instance of Plasma.LineShiftScheme.

JAC.Plasma.PartialWaveData — Type

struct Plasma.PartialWaveData ... defines a type to collect for partial-waves (kappa) cross sections, rates, etc. at different energies

+ kappa    ::Int64                            ... kappa
+ pairs    ::Vector{Tuple{Float64, Float64}}  
    ... vector of pairs, for instance (energy, cs), to later combine data in a proper manner.

JAC.Plasma.PartialWaveData — Method

Plasma.PartialWaveData() ... constructor for the default values of Plasma.PartialWaveData set

JAC.Plasma.SahaBoltzmannScheme — Type

struct Plasma.SahaBoltzmannScheme <: Plasma.AbstractPlasmaScheme ... a struct to thermodynamic properties of a Saha-Boltzmann LTE mixture..

+ plasmaModel           ::Basics.AbstractPlasmaModel                 
    ... A plasma model that "restricts" the Saha-Boltzmann equilibrium densities by some additional parameters, for instance,
        due to ionization-potential-depression (IPD) or others.
+ calcLTE               ::Bool                 ... True, if the Saha-Boltzmann equilibrium densities should be calculated.         
+ printIonLevels        ::Bool                 ... True, for printing detailed information about all ionic levels.         
+ qRange                ::UnitRange{Int64}     ... Range of charge states q for which ions are taken into account. 
+ maxNoIonLevels        ::Int64                ... (maximum) No of ionic levels for any charge state of the ions in the mixture.
+ NoExcitations         ::Int64                
    ... No of excitations (S, D, T) that are taken into account with regard to the reference (ground) configuration of the ions.
        This number is taken as a second qualifier to characterize the quality of the ionic-level data. Usually, NoExcitations = 1,2.
+ upperShellNo          ::Int64                
    ... upper-most princicpal quantum number n for which orbitals are taken into account into the ionic-level computations;
        this is often chosen upperShellNo= 4..10; if ionic-level data are given, this number decides which of the levels
        are taken into account.
+ isotopicMixture       ::Array{Basics.IsotopicFraction,1}   
    ... List of (non-normlized) isotopic fractions that form the requested mixture; the fractions will first be renormalized
        to 1 in course of the Saha-Boltzmann LTE computations.
+ isotopeFilenames      ::Array{String,1}     
    ... set of files names from which the ionic-level data can be read in for the different isotopes (Z,A) in the mixture.

JAC.Plasma.SahaBoltzmannScheme — Method

Plasma.SahaBoltzmannScheme() ... constructor for an 'default' instance of a Plasma.SahaBoltzmannScheme.

JAC.Plasma.Settings — Type

struct Plasma.Settings ... defines a type for the details and parameters of computing photoionization lines.

+ temperature               ::Float64     ... Plasma temperature in [K].
+ density                   ::Float64     ... Plasma density in [g/cm^3].
+ useNumberDensity          ::Bool    
    ... true, if the density above is taken as (total ion) number density ni, and false otherwise.

JAC.Plasma.Settings — Method

Plasma.Settings(set::Plasma.Settings;

    temperature=..,         density=..,         useNumberDensity =..)
                
... constructor for modifying the given Plasma.Settings by 'overwriting' the previously selected parameters.

JAC.Plasma.Settings — Method

Plasma.Settings() ... constructor for the default values of plasma computations

JAC.Basics.perform — Method

Basics.perform(comp::Plasma.Computation) ... to set-up and perform a plasma computation that starts from a given set of reference configurations and support both, an atomic-average SCF procedure and the computation of various plasma properties and processe. The results of all individual steps are printed to screen but nothing is returned otherwise.

Basics.perform(comp::Plasma.Computation; output::Bool=true) ... to perform the same but to return the complete output in a dictionary; the particular output depends on the type and specifications of the plasma computation but can easily accessed by the keys of this dictionary.

Empirical plasma rates

Line-shift computations

Saha-Boltzmann computations of ionic mixtures